+++ Sources/VaporWebsite/Translations/BrazilianPortuguese.swift
@@ -1,0 +1,102 @@
+// BrazilianPortuguese strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let brazilianPortugueseStrings: [String: String] = [
+    "home.hero.title": "Swift, mas no servidor",
+    "home.hero.caption": "O Vapor oferece uma base segura, eficiente e fácil de usar para criar servidores HTTP, backends e APIs em Swift",
+    // Buttons / controls.
+    "cta.getStarted": "Começar",
+    "cta.learnMore": "Saiba mais",
+    "cta.seeInAction": "Veja em ação",
+    "home.apiDocs.goToApiDocs": "Ir para a documentação da API",
+    "home.discord.contribute": "Contribuir",
+    "home.showcase.seeFullList": "Ver lista completa",
+    "home.showcase.prev": "Showcase anterior",
+    "home.showcase.next": "Próximo showcase",
+    "home.sponsors.becomeSupporter": "Torne-se um apoiador",
+    "home.sponsors.prev": "Patrocinadores anteriores",
+    "home.sponsors.next": "Próximos patrocinadores",
+    "showcase.title": "O Vapor impulsiona milhares de apps e sites de alto desempenho",
+    "showcase.caption": "Conheça os apps, sites e empresas que constroem com o Vapor — e os times por trás deles.",
+    "showcase.search": "Buscar",
+    "showcase.searchLabel": "Buscar na vitrine",
+    "showcase.sortLabel": "Ordenar a vitrine",
+    "showcase.sort.newest": "Mais recentes primeiro",
+    "showcase.sort.oldest": "Mais antigos primeiro",
+    "showcase.sort.nameAsc": "Nome (A–Z)",
+    "showcase.sort.nameDesc": "Nome (Z–A)",
+    "showcase.empty": "Nenhum app corresponde à sua busca.",
+    "showcase.paginationLabel": "Paginação da vitrine",
+    "showcase.cta.title": "Usa o Vapor no seu app? Adoraríamos destacá-lo!",
+    "showcase.cta.caption": "Ajude a mostrar as diversas formas de usar o Vapor.",
+    "showcase.cta.button": "Entre em contato",
+    "footer.joinDiscord": "Entre no nosso Discord",
+    // Nav / footer chrome.
+    "nav.documentation": "Documentação",
+    "nav.frameworkDocs": "Documentação do framework",
+    "nav.frameworkDocs.caption": "Aprenda a usar o Vapor",
+    "nav.apiDocs": "Documentação da API",
+    "nav.apiDocs.caption": "Documentação de referência do Vapor",
+    "nav.selectTheme": "Selecionar tema",
+    "nav.closeMenu": "Fechar menu",
+    "footer.supporters": "Apoiadores",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} estrelas no GitHub",
+    "home.hero.starsLoading": "… estrelas no GitHub",
+    "home.companies.caption": "Impulsionando empresas como:",
+    "home.packages.title": "Crie o lado do servidor com as ferramentas que você já conhece",
+    "home.packages.caption": "Não é preciso aprender uma linguagem do zero (nem montar uma equipe separada) só para o seu backend. O Vapor é construído sobre o SwiftNIO da Apple, então você poderá usar a linguagem que você já conhece e ama.",
+    "home.packages.vapor.description": "Crie APIs eficientes em uma linguagem que você ama. Defina rotas, envie e receba JSON e construa servidores HTTP.",
+    "home.packages.fluent.description": "Crie modelos e interaja com seu banco de dados em código Swift nativo e seguro, sem precisar escrever nenhum SQL.",
+    "home.packages.jwt.description": "Crie, assine e verifique JSON Web Tokens em Swift. Construído sobre o Swift Crypto.",
+    "home.packages.leaf.description": "Um motor de templates escrito em Swift. Gere HTML para aplicações web e e-mails com uma sintaxe simples que qualquer pessoa pode usar.",
+    "home.features.performance.title": "APIs e servidores de alto desempenho",
+    "home.features.performance.description": "Construído com uma arquitetura não bloqueante e orientada a eventos, o Vapor permite criar APIs e servidores HTTP escaláveis e de alto desempenho. Com o modelo de concorrência do Swift, você pode escrever um código claro e fácil de manter, eficiente e legível.",
+    "home.features.confidence.title": "Faça deploy com confiança, mesmo às sextas-feiras",
+    "home.features.confidence.description": "Com o design expressivo e orientado a protocolos do Vapor, você terá tranquilidade ao publicar seu código. Com nosso forte foco em segurança de tipos, muitos erros e problemas são detectados cedo pelo compilador.",
+    "home.integrations.label": "Integrações",
+    "home.integrations.title": "Integração completa com toda a sua stack",
+    "home.integrations.description": "O ecossistema maduro do Vapor inclui mais de cem pacotes Swift voltados para o servidor, oficiais e mantidos pela comunidade, para tornar o desenvolvimento das suas aplicações fácil e eficiente.",
+    "home.showcase.title": "Usado pelas equipes por trás destes aplicativos incríveis",
+    "home.sponsors.title": "Apoiado por nossos incríveis patrocinadores e colaboradores",
+    "home.discord.title": "Junte-se à maior comunidade de desenvolvedores backend de Swift",
+    "home.discord.description": "A comunidade do Discord do Vapor, com mais de 13 mil membros, estará ao seu lado para ajudar você ao longo do caminho. Faça perguntas, contribua e faça parte de um cantinho próspero e acolhedor da internet.",
+    "home.apiDocs.title": "Obtenha todas as ferramentas necessárias para desenvolver com Swift",
+    "home.apiDocs.item1": "Crie novos projetos facilmente com o Vapor Toolbox",
+    "home.apiDocs.item2": "Documentação abrangente e referência da API",
+    "home.apiDocs.item3": "Tudo o que você precisa para criar backends e APIs",
+    "home.apiDocs.item4": "Suporte completo ao modelo de concorrência do Swift",
+    "comingSoon.message": "Esta página ainda está em construção. Fique de olho nas novidades!",
+    "pages.showcase.title": "Vitrine",
+    "pages.team.title": "Equipe",
+    "pages.supporters.title": "Apoiadores",
+    "pages.evangelists.title": "Evangelistas",
+]
+
+let brazilianPortugueseLanguage = Language(
+    .brazilianPortuguese,
+    customStrings: brazilianPortugueseStrings,
+    image: "static/images/opengraph/vapor-og-pt-BR-2x.png",
+    localisation: .init(
+        notFoundTitle: "Página não encontrada",
+        notFoundMessage: "A página que você procura pode ter sido movida, renomeada ou nunca ter existido.",
+        notFoundLink: "Voltar para a página inicial",
+        toggleNavigation: "Alternar navegação",
+        skipToContent: "Pular para o conteúdo",
+        home: "Início",
+        store: "Loja",
+        blog: "Blog",
+        showcase: "Vitrine",
+        team: "Equipe",
+        help: "Ajuda",
+        pressKit: "Kit de imprensa",
+        community: "Comunidade",
+        resources: "Recursos",
+        language: "Idioma",
+        theme: "Tema",
+        lightTheme: "Claro",
+        darkTheme: "Escuro",
+        systemTheme: "Sistema"
+    )
+)
