+++ Sources/VaporWebsite/Translations/Chinese.swift
@@ -1,0 +1,102 @@
+// Chinese strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let chineseStrings: [String: String] = [
+    "home.hero.title": "Swift，但在服务器上",
+    "home.hero.caption": "Vapor 提供了一个安全、高性能且易于使用的基础，用于在 Swift 中构建 HTTP 服务器、后端和 API",
+    // Buttons / controls.
+    "cta.getStarted": "开始使用",
+    "cta.learnMore": "了解更多",
+    "cta.seeInAction": "查看实际效果",
+    "home.apiDocs.goToApiDocs": "前往 API 文档",
+    "home.discord.contribute": "参与贡献",
+    "home.showcase.seeFullList": "查看完整列表",
+    "home.showcase.prev": "上一个展示",
+    "home.showcase.next": "下一个展示",
+    "home.sponsors.becomeSupporter": "成为赞助者",
+    "home.sponsors.prev": "上一批赞助者",
+    "home.sponsors.next": "下一批赞助者",
+    "showcase.title": "Vapor 为数千款高性能应用和网站提供支持",
+    "showcase.caption": "探索使用 Vapor 构建的应用、网站和公司，以及背后的团队。",
+    "showcase.search": "搜索",
+    "showcase.searchLabel": "搜索案例展示",
+    "showcase.sortLabel": "排序案例展示",
+    "showcase.sort.newest": "最新优先",
+    "showcase.sort.oldest": "最早优先",
+    "showcase.sort.nameAsc": "名称（A–Z）",
+    "showcase.sort.nameDesc": "名称（Z–A）",
+    "showcase.empty": "没有符合搜索条件的应用。",
+    "showcase.paginationLabel": "案例展示分页",
+    "showcase.cta.title": "在你的应用中使用了 Vapor 吗？我们很乐意展示它！",
+    "showcase.cta.caption": "帮助展示 Vapor 的各种使用方式。",
+    "showcase.cta.button": "联系我们",
+    "footer.joinDiscord": "加入我们的 Discord",
+    // Nav / footer chrome.
+    "nav.documentation": "文档",
+    "nav.frameworkDocs": "框架文档",
+    "nav.frameworkDocs.caption": "学习如何使用 Vapor",
+    "nav.apiDocs": "API 文档",
+    "nav.apiDocs.caption": "Vapor 参考文档",
+    "nav.selectTheme": "选择主题",
+    "nav.closeMenu": "关闭菜单",
+    "footer.supporters": "赞助者",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} 个 GitHub star",
+    "home.hero.starsLoading": "… 个 GitHub star",
+    "home.companies.caption": "为以下公司提供支持：",
+    "home.packages.title": "使用你已熟悉的工具构建服务器端",
+    "home.packages.caption": "无需仅仅为了后端而从头学习一门语言（或组建另一支团队）。Vapor 基于 Apple 的 SwiftNIO 构建，因此你可以继续使用你已经熟悉和喜爱的语言。",
+    "home.packages.vapor.description": "用你喜爱的语言构建高效的 API。创建路由、收发 JSON 并构建 HTTP 服务器。",
+    "home.packages.fluent.description": "使用原生、安全的 Swift 代码创建模型并操作数据库，无需编写任何 SQL。",
+    "home.packages.jwt.description": "在 Swift 中创建、签名和验证 JSON Web Token。基于 Swift Crypto 构建。",
+    "home.packages.leaf.description": "用 Swift 编写的模板引擎。使用任何人都能上手的简单语法，为 Web 应用和电子邮件生成 HTML。",
+    "home.features.performance.title": "高性能的 API 和服务器",
+    "home.features.performance.description": "Vapor 采用非阻塞、事件驱动的架构，让你能够构建高性能、可扩展的 API 和 HTTP 服务器。借助 Swift 的并发模型，你可以编写清晰、易维护、高效且易读的代码。",
+    "home.features.confidence.title": "即使在周五也能放心发布",
+    "home.features.confidence.description": "凭借 Vapor 富有表现力、面向协议的设计，你在发布代码时会更加安心。由于我们高度重视类型安全，许多错误和问题在编译阶段就能被及早发现。",
+    "home.integrations.label": "集成",
+    "home.integrations.title": "与你的整个技术栈完全集成",
+    "home.integrations.description": "Vapor 成熟的生态系统包含一百多个官方及社区维护的服务器优先 Swift 包，让你的应用开发更轻松、更高效。",
+    "home.showcase.title": "这些出色应用背后的团队都在使用",
+    "home.sponsors.title": "由我们了不起的赞助者和支持者提供支持",
+    "home.discord.title": "加入最大的 Swift 后端开发者社区",
+    "home.discord.description": "Vapor 拥有超过 13,000 名成员的 Discord 社区将一路为你提供支持。提出问题、参与贡献，成为这个充满活力、友善的网络角落的一员。",
+    "home.apiDocs.title": "获取使用 Swift 构建所需的全部工具",
+    "home.apiDocs.item1": "使用 Vapor Toolbox 轻松创建新项目",
+    "home.apiDocs.item2": "丰富的文档和 API 参考",
+    "home.apiDocs.item3": "构建后端和 API 所需的一切",
+    "home.apiDocs.item4": "全面支持 Swift 的并发模型",
+    "comingSoon.message": "此页面仍在建设中，敬请期待更新！",
+    "pages.showcase.title": "案例展示",
+    "pages.team.title": "团队",
+    "pages.supporters.title": "赞助者",
+    "pages.evangelists.title": "布道师",
+]
+
+let chineseLanguage = Language(
+    .chinese,
+    customStrings: chineseStrings,
+    image: "static/images/opengraph/vapor-og-zh-2x.png",
+    localisation: .init(
+        notFoundTitle: "未找到页面",
+        notFoundMessage: "您要查找的页面可能已被移动、重命名或从未存在。",
+        notFoundLink: "返回首页",
+        toggleNavigation: "切换导航",
+        skipToContent: "跳到内容",
+        home: "首页",
+        store: "商店",
+        blog: "博客",
+        showcase: "案例展示",
+        team: "团队",
+        help: "帮助",
+        pressKit: "媒体资源",
+        community: "社区",
+        resources: "资源",
+        language: "语言",
+        theme: "主题",
+        lightTheme: "浅色",
+        darkTheme: "深色",
+        systemTheme: "跟随系统"
+    )
+)
