+++ Sources/VaporWebsite/Translations/Dutch.swift
@@ -1,0 +1,102 @@
+// Dutch strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let dutchStrings: [String: String] = [
+    "home.hero.title": "Swift, maar dan op een server",
+    "home.hero.caption": "Vapor biedt een veilige, performante en gebruiksvriendelijke basis om HTTP-servers, backends en API's in Swift te bouwen",
+    // Buttons / controls.
+    "cta.getStarted": "Aan de slag",
+    "cta.learnMore": "Meer informatie",
+    "cta.seeInAction": "Zie het in actie",
+    "home.apiDocs.goToApiDocs": "Naar de API-documentatie",
+    "home.discord.contribute": "Bijdragen",
+    "home.showcase.seeFullList": "Volledige lijst bekijken",
+    "home.showcase.prev": "Vorige showcase",
+    "home.showcase.next": "Volgende showcase",
+    "home.sponsors.becomeSupporter": "Word ondersteuner",
+    "home.sponsors.prev": "Vorige sponsors",
+    "home.sponsors.next": "Volgende sponsors",
+    "showcase.title": "Vapor drijft duizenden goed presterende apps en sites aan",
+    "showcase.caption": "Ontdek de apps, sites en bedrijven die op Vapor bouwen – en de teams erachter.",
+    "showcase.search": "Zoeken",
+    "showcase.searchLabel": "Etalage doorzoeken",
+    "showcase.sortLabel": "Etalage sorteren",
+    "showcase.sort.newest": "Nieuwste eerst",
+    "showcase.sort.oldest": "Oudste eerst",
+    "showcase.sort.nameAsc": "Naam (A–Z)",
+    "showcase.sort.nameDesc": "Naam (Z–A)",
+    "showcase.empty": "Geen apps komen overeen met je zoekopdracht.",
+    "showcase.paginationLabel": "Etalagepaginering",
+    "showcase.cta.title": "Gebruik je Vapor in je app? We laten je graag zien!",
+    "showcase.cta.caption": "Help laten zien op hoeveel verschillende manieren Vapor gebruikt kan worden.",
+    "showcase.cta.button": "Neem contact op",
+    "footer.joinDiscord": "Word lid van onze Discord",
+    // Nav / footer chrome.
+    "nav.documentation": "Documentatie",
+    "nav.frameworkDocs": "Framework-documentatie",
+    "nav.frameworkDocs.caption": "Leer hoe je Vapor gebruikt",
+    "nav.apiDocs": "API-documentatie",
+    "nav.apiDocs.caption": "Referentiedocumentatie voor Vapor",
+    "nav.selectTheme": "Selecteer thema",
+    "nav.closeMenu": "Menu sluiten",
+    "footer.supporters": "Ondersteuners",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} sterren op GitHub",
+    "home.hero.starsLoading": "… sterren op GitHub",
+    "home.companies.caption": "Drijft bedrijven aan zoals:",
+    "home.packages.title": "Bouw server-side met de tools die je al kent",
+    "home.packages.caption": "Je hoeft geen taal vanaf nul te leren (of een apart team samen te stellen) alleen voor je backend. Vapor is gebouwd op Apple’s SwiftNIO, dus je kunt de taal gebruiken die je al kent en waardeert.",
+    "home.packages.vapor.description": "Bouw efficiënte API's in een taal die je geweldig vindt. Maak routes, verstuur en ontvang JSON en bouw HTTP-servers.",
+    "home.packages.fluent.description": "Maak modellen en werk met je database in native, veilige Swift-code zonder ook maar één regel SQL te schrijven.",
+    "home.packages.jwt.description": "Maak, onderteken en verifieer JSON Web Tokens in Swift. Gebouwd op Swift Crypto.",
+    "home.packages.leaf.description": "Een template-engine geschreven in Swift. Genereer HTML voor zowel webapps als e-mails met een eenvoudige syntaxis die iedereen kan gebruiken.",
+    "home.features.performance.title": "Hoogperformante API's en servers",
+    "home.features.performance.description": "Vapor is gebouwd op een non-blocking, event-driven architectuur en stelt je in staat hoogperformante, schaalbare API's en HTTP-servers te bouwen. Met het Concurrency-model van Swift schrijf je heldere, onderhoudbare code die efficiënt en gemakkelijk te lezen is.",
+    "home.features.confidence.title": "Lever met vertrouwen op, zelfs op vrijdag",
+    "home.features.confidence.description": "Met het expressieve, protocol-georiënteerde ontwerp van Vapor lever je je code met een gerust hart op. Door onze sterke focus op type-veiligheid worden veel fouten en problemen al vroeg door de compiler opgespoord.",
+    "home.integrations.label": "Integraties",
+    "home.integrations.title": "Volledige integratie met je hele stack",
+    "home.integrations.description": "Het volwassen ecosysteem van Vapor omvat meer dan honderd officiële en door de community onderhouden server-first Swift-packages om het bouwen van je applicaties eenvoudig en efficiënt te maken.",
+    "home.showcase.title": "Gebruikt door de teams achter deze geweldige apps",
+    "home.sponsors.title": "Ondersteund door onze geweldige sponsors en donateurs",
+    "home.discord.title": "Word lid van de grootste community van Swift-backendontwikkelaars",
+    "home.discord.description": "De Discord-community van Vapor met meer dan 13.000 leden staat je bij op je weg. Stel vragen, draag bij en word deel van een bloeiend, hartelijk hoekje van het internet.",
+    "home.apiDocs.title": "Krijg alle tools die je nodig hebt om met Swift te bouwen",
+    "home.apiDocs.item1": "Maak eenvoudig nieuwe projecten met de Vapor Toolbox",
+    "home.apiDocs.item2": "Uitgebreide documentatie en API-referentie",
+    "home.apiDocs.item3": "Alles wat je nodig hebt om backends en API's te bouwen",
+    "home.apiDocs.item4": "Volledige ondersteuning voor het concurrency-model van Swift",
+    "comingSoon.message": "Deze pagina is nog in aanbouw, houd updates in de gaten!",
+    "pages.showcase.title": "Etalage",
+    "pages.team.title": "Team",
+    "pages.supporters.title": "Ondersteuners",
+    "pages.evangelists.title": "Evangelisten",
+]
+
+let dutchLanguage = Language(
+    .dutch,
+    customStrings: dutchStrings,
+    image: "static/images/opengraph/vapor-og-nl-2x.png",
+    localisation: .init(
+        notFoundTitle: "Pagina niet gevonden",
+        notFoundMessage: "De pagina die je zoekt is mogelijk verplaatst, hernoemd of heeft nooit bestaan.",
+        notFoundLink: "Terug naar de startpagina",
+        toggleNavigation: "Navigatie schakelen",
+        skipToContent: "Naar de inhoud",
+        home: "Home",
+        store: "Winkel",
+        blog: "Blog",
+        showcase: "Etalage",
+        team: "Team",
+        help: "Help",
+        pressKit: "Perskit",
+        community: "Community",
+        resources: "Bronnen",
+        language: "Taal",
+        theme: "Thema",
+        lightTheme: "Licht",
+        darkTheme: "Donker",
+        systemTheme: "Systeem"
+    )
+)
