+++ Sources/VaporWebsite/Translations/French.swift
@@ -1,0 +1,102 @@
+// French strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let frenchStrings: [String: String] = [
+    "home.hero.title": "Swift, mais sur le serveur",
+    "home.hero.caption": "Vapor fournit une base sûre, performante et facile à utiliser pour créer des serveurs HTTP, des backends et des API en Swift",
+    // Buttons / controls.
+    "cta.getStarted": "Commencer",
+    "cta.learnMore": "En savoir plus",
+    "cta.seeInAction": "Voir en action",
+    "home.apiDocs.goToApiDocs": "Accéder à la documentation de l’API",
+    "home.discord.contribute": "Contribuer",
+    "home.showcase.seeFullList": "Voir la liste complète",
+    "home.showcase.prev": "Vitrine précédente",
+    "home.showcase.next": "Vitrine suivante",
+    "home.sponsors.becomeSupporter": "Devenir soutien",
+    "home.sponsors.prev": "Sponsors précédents",
+    "home.sponsors.next": "Sponsors suivants",
+    "showcase.title": "Vapor propulse des milliers d'applications et de sites performants",
+    "showcase.caption": "Découvrez les applications, les sites et les entreprises qui s'appuient sur Vapor, et les équipes derrière eux.",
+    "showcase.search": "Rechercher",
+    "showcase.searchLabel": "Rechercher dans la vitrine",
+    "showcase.sortLabel": "Trier la vitrine",
+    "showcase.sort.newest": "Plus récents d'abord",
+    "showcase.sort.oldest": "Plus anciens d'abord",
+    "showcase.sort.nameAsc": "Nom (A–Z)",
+    "showcase.sort.nameDesc": "Nom (Z–A)",
+    "showcase.empty": "Aucune application ne correspond à votre recherche.",
+    "showcase.paginationLabel": "Pagination de la vitrine",
+    "showcase.cta.title": "Vous utilisez Vapor dans votre application ? Nous serions ravis de la mettre en avant !",
+    "showcase.cta.caption": "Aidez à montrer les différentes façons d'utiliser Vapor.",
+    "showcase.cta.button": "Nous contacter",
+    "footer.joinDiscord": "Rejoignez notre Discord",
+    // Nav / footer chrome.
+    "nav.documentation": "Documentation",
+    "nav.frameworkDocs": "Documentation du framework",
+    "nav.frameworkDocs.caption": "Apprenez à utiliser Vapor",
+    "nav.apiDocs": "Documentation de l’API",
+    "nav.apiDocs.caption": "Documentation de référence de Vapor",
+    "nav.selectTheme": "Sélectionner le thème",
+    "nav.closeMenu": "Fermer le menu",
+    "footer.supporters": "Soutiens",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} étoiles sur GitHub",
+    "home.hero.starsLoading": "… étoiles sur GitHub",
+    "home.companies.caption": "Utilisé par des entreprises comme :",
+    "home.packages.title": "Développez côté serveur avec les outils que vous maîtrisez déjà",
+    "home.packages.caption": "Pas besoin d’apprendre un langage à partir de zéro (ni de constituer une autre équipe) juste pour votre backend. Vapor repose sur SwiftNIO d’Apple, vous pourrez donc utiliser le langage que vous connaissez et appréciez déjà.",
+    "home.packages.vapor.description": "Créez des API efficaces dans un langage que vous aimez. Définissez des routes, envoyez et recevez du JSON et créez des serveurs HTTP.",
+    "home.packages.fluent.description": "Créez des modèles et interagissez avec votre base de données en code Swift natif et sûr, sans avoir à écrire de SQL.",
+    "home.packages.jwt.description": "Créez, signez et vérifiez des JSON Web Tokens en Swift. Basé sur Swift Crypto.",
+    "home.packages.leaf.description": "Un moteur de templates écrit en Swift. Générez du HTML pour vos applications web et vos e-mails avec une syntaxe simple, à la portée de tous.",
+    "home.features.performance.title": "Des API et des serveurs hautement performants",
+    "home.features.performance.description": "Bâti sur une architecture non bloquante et événementielle, Vapor vous permet de créer des API et des serveurs HTTP performants et évolutifs. Grâce au modèle de concurrence de Swift, vous écrivez un code clair et maintenable, efficace et facile à lire.",
+    "home.features.confidence.title": "Déployez en toute confiance, même le vendredi",
+    "home.features.confidence.description": "Grâce à la conception expressive et orientée protocole de Vapor, vous déployez votre code l’esprit tranquille. Avec notre fort accent sur la sûreté du typage, de nombreuses erreurs et problèmes sont détectés très tôt par le compilateur.",
+    "home.integrations.label": "Intégrations",
+    "home.integrations.title": "Une intégration complète avec l’ensemble de votre stack",
+    "home.integrations.description": "L’écosystème mature de Vapor comprend plus d’une centaine de paquets Swift orientés serveur, officiels et maintenus par la communauté, pour rendre le développement de vos applications simple et efficace.",
+    "home.showcase.title": "Utilisé par les équipes derrière ces applications incroyables",
+    "home.sponsors.title": "Soutenu par nos incroyables sponsors et donateurs",
+    "home.discord.title": "Rejoignez la plus grande communauté de développeurs backend Swift",
+    "home.discord.description": "La communauté Discord de Vapor, forte de plus de 13 000 membres, sera à vos côtés pour vous accompagner. Posez des questions, contribuez et faites partie d’un coin d’internet dynamique et bienveillant.",
+    "home.apiDocs.title": "Obtenez tous les outils dont vous avez besoin pour développer avec Swift",
+    "home.apiDocs.item1": "Créez facilement de nouveaux projets avec la Vapor Toolbox",
+    "home.apiDocs.item2": "Une documentation complète et une référence d’API",
+    "home.apiDocs.item3": "Tout ce dont vous avez besoin pour créer des backends et des API",
+    "home.apiDocs.item4": "Une prise en charge complète du modèle de concurrence de Swift",
+    "comingSoon.message": "Cette page est encore en construction, restez à l’écoute pour les mises à jour !",
+    "pages.showcase.title": "Vitrine",
+    "pages.team.title": "Équipe",
+    "pages.supporters.title": "Soutiens",
+    "pages.evangelists.title": "Évangélistes",
+]
+
+let frenchLanguage = Language(
+    .french,
+    customStrings: frenchStrings,
+    image: "static/images/opengraph/vapor-og-fr-2x.png",
+    localisation: .init(
+        notFoundTitle: "Page introuvable",
+        notFoundMessage: "La page que vous recherchez a peut-être été déplacée, renommée ou n’a jamais existé.",
+        notFoundLink: "Retour à la page d’accueil",
+        toggleNavigation: "Afficher/masquer la navigation",
+        skipToContent: "Passer au contenu",
+        home: "Accueil",
+        store: "Boutique",
+        blog: "Blog",
+        showcase: "Vitrine",
+        team: "Équipe",
+        help: "Aide",
+        pressKit: "Kit de presse",
+        community: "Communauté",
+        resources: "Ressources",
+        language: "Langue",
+        theme: "Thème",
+        lightTheme: "Clair",
+        darkTheme: "Sombre",
+        systemTheme: "Système"
+    )
+)
