+++ Sources/VaporWebsite/Translations/German.swift
@@ -1,0 +1,99 @@
+// German strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let germanStrings: [String: String] = [
+    "nav.closeMenu": "Menü schließen",
+    "nav.documentation": "Dokumentation",
+    "nav.frameworkDocs": "Framework-Dokumentation",
+    "nav.frameworkDocs.caption": "Erfahren Sie, wie Sie Vapor verwenden",
+    "nav.apiDocs": "API-Dokumentation",
+    "nav.apiDocs.caption": "Referenzdokumentation für Vapor",
+    "nav.selectTheme": "Thema auswählen",
+    "footer.joinDiscord": "Tritt unserem Discord bei",
+    "footer.supporters": "Unterstützer",
+    "comingSoon.message": "Diese Seite befindet sich noch im Aufbau – bleib dran für Updates!",
+    "pages.showcase.title": "Showcase",
+    "pages.team.title": "Team",
+    "pages.supporters.title": "Unterstützer",
+    "pages.evangelists.title": "Evangelisten",
+    "cta.getStarted": "Loslegen",
+    "cta.learnMore": "Mehr erfahren",
+    "cta.seeInAction": "In Aktion sehen",
+    "home.hero.title": "Swift, aber auf dem Server",
+    "home.hero.caption": "Vapor bietet eine sichere, performante und einfach zu bedienende Grundlage, um HTTP-Server, Backends und APIs in Swift zu erstellen",
+    "home.hero.stars": "{count} Sterne auf GitHub",
+    "home.hero.starsLoading": "… Sterne auf GitHub",
+    "home.companies.caption": "Im Einsatz bei Unternehmen wie:",
+    "home.packages.title": "Entwickle serverseitig mit den Werkzeugen, die du bereits kennst",
+    "home.packages.caption": "Du musst keine Sprache von Grund auf neu lernen (oder ein eigenes Team aufbauen) nur für dein Backend. Vapor baut auf Apples SwiftNIO auf, sodass du die Sprache verwenden kannst, die du bereits kennst und liebst.",
+    "home.packages.vapor.description": "Erstelle effiziente APIs in einer Sprache, die du liebst. Definiere Routen, sende und empfange JSON und baue HTTP-Server.",
+    "home.packages.fluent.description": "Erstelle Models und interagiere mit deiner Datenbank in nativem, sicherem Swift-Code, ohne SQL schreiben zu müssen.",
+    "home.packages.jwt.description": "Erstelle, signiere und verifiziere JSON Web Tokens in Swift. Basiert auf Swift Crypto.",
+    "home.packages.leaf.description": "Eine in Swift geschriebene Template-Engine. Generiere HTML für Web-Apps und E-Mails mit einer einfachen Syntax, die jeder nutzen kann.",
+    "home.features.performance.title": "Hochperformante APIs und Server",
+    "home.features.performance.description": "Dank einer nicht-blockierenden, ereignisgesteuerten Architektur kannst du mit Vapor hochperformante, skalierbare APIs und HTTP-Server erstellen. Mit dem Concurrency-Modell von Swift schreibst du klaren, wartbaren Code, der effizient und leicht zu lesen ist.",
+    "home.features.confidence.title": "Veröffentliche mit Zuversicht – sogar freitags",
+    "home.features.confidence.description": "Dank Vapors ausdrucksstarkem, protokollorientiertem Design kannst du deinen Code mit gutem Gewissen veröffentlichen. Durch unseren starken Fokus auf Typsicherheit werden viele Fehler und Probleme bereits früh vom Compiler erkannt.",
+    "home.integrations.label": "Integrationen",
+    "home.integrations.title": "Vollständige Integration in deinen gesamten Stack",
+    "home.integrations.description": "Vapors ausgereiftes Ökosystem umfasst über hundert offizielle und von der Community gepflegte, serverorientierte Swift-Pakete, die das Entwickeln deiner Anwendungen einfach und effizient machen.",
+    "home.showcase.title": "Verwendet von den Teams hinter diesen großartigen Apps",
+    "home.discord.title": "Tritt der größten Community von Swift-Backend-Entwicklern bei",
+    "home.discord.description": "Vapors über 13.000 Mitglieder zählende Discord-Community steht dir auf deinem Weg zur Seite. Stelle Fragen, bring dich ein und werde Teil einer lebendigen, freundlichen Ecke des Internets.",
+    "home.apiDocs.title": "Erhalte alle Werkzeuge, die du zum Entwickeln mit Swift brauchst",
+    "home.apiDocs.item1": "Erstelle mühelos neue Projekte mit der Vapor Toolbox",
+    "home.apiDocs.item2": "Umfangreiche Dokumentation und API-Referenz",
+    "home.apiDocs.item3": "Alles, was du zum Entwickeln von Backends und APIs brauchst",
+    "home.apiDocs.item4": "Volle Unterstützung für das Concurrency-Modell von Swift",
+    "home.sponsors.title": "Unterstützt von unseren großartigen Sponsoren und Förderern",
+    "home.sponsors.becomeSupporter": "Werde Unterstützer",
+    "home.apiDocs.goToApiDocs": "Zur API-Dokumentation",
+    "home.discord.contribute": "Mitwirken",
+    "home.showcase.seeFullList": "Vollständige Liste ansehen",
+    "home.showcase.prev": "Vorheriges Showcase",
+    "home.showcase.next": "Nächstes Showcase",
+    "home.sponsors.prev": "Vorherige Sponsoren",
+    "home.sponsors.next": "Nächste Sponsoren",
+    "showcase.title": "Vapor treibt Tausende leistungsstarke Apps und Websites an",
+    "showcase.caption": "Entdecke die Apps, Websites und Unternehmen, die auf Vapor setzen – und die Teams dahinter.",
+    "showcase.search": "Suchen",
+    "showcase.searchLabel": "Showcase durchsuchen",
+    "showcase.sortLabel": "Showcase sortieren",
+    "showcase.sort.newest": "Neueste zuerst",
+    "showcase.sort.oldest": "Älteste zuerst",
+    "showcase.sort.nameAsc": "Name (A–Z)",
+    "showcase.sort.nameDesc": "Name (Z–A)",
+    "showcase.empty": "Keine Apps entsprechen deiner Suche.",
+    "showcase.paginationLabel": "Showcase-Seitennavigation",
+    "showcase.cta.title": "Nutzt du Vapor in deiner App? Wir würden dich gerne vorstellen!",
+    "showcase.cta.caption": "Hilf zu zeigen, auf wie viele verschiedene Arten Vapor eingesetzt werden kann.",
+    "showcase.cta.button": "Kontakt aufnehmen",
+]
+
+let germanLanguage = Language(
+    .german,
+    customStrings: germanStrings,
+    image: "static/images/opengraph/vapor-og-de-2x.png",
+    localisation: .init(
+        notFoundTitle: "Seite nicht gefunden",
+        notFoundMessage: "Die gesuchte Seite wurde möglicherweise verschoben, umbenannt oder existiert nicht.",
+        notFoundLink: "Zurück zur Startseite",
+        toggleNavigation: "Navigation umschalten",
+        skipToContent: "Zum Inhalt springen",
+        home: "Startseite",
+        store: "Shop",
+        blog: "Blog",
+        showcase: "Showcase",
+        team: "Team",
+        help: "Hilfe",
+        pressKit: "Pressekit",
+        community: "Community",
+        resources: "Ressourcen",
+        language: "Sprache",
+        theme: "Thema",
+        lightTheme: "Hell",
+        darkTheme: "Dunkel",
+        systemTheme: "System"
+    )
+)
