+++ Sources/VaporWebsite/Translations/Italian.swift
@@ -1,0 +1,102 @@
+// Italian strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let italianStrings: [String: String] = [
+    "home.hero.title": "Swift, ma sul server",
+    "home.hero.caption": "Vapor offre una base sicura, performante e facile da usare per creare server HTTP, backend e API in Swift",
+    // Buttons / controls.
+    "cta.getStarted": "Inizia",
+    "cta.learnMore": "Scopri di più",
+    "cta.seeInAction": "Guardalo in azione",
+    "home.apiDocs.goToApiDocs": "Vai alla documentazione API",
+    "home.discord.contribute": "Contribuisci",
+    "home.showcase.seeFullList": "Vedi l’elenco completo",
+    "home.showcase.prev": "Showcase precedente",
+    "home.showcase.next": "Showcase successivo",
+    "home.sponsors.becomeSupporter": "Diventa sostenitore",
+    "home.sponsors.prev": "Sponsor precedenti",
+    "home.sponsors.next": "Sponsor successivi",
+    "showcase.title": "Vapor alimenta migliaia di app e siti ad alte prestazioni",
+    "showcase.caption": "Scopri le app, i siti e le aziende che si basano su Vapor, e i team dietro di essi.",
+    "showcase.search": "Cerca",
+    "showcase.searchLabel": "Cerca nella vetrina",
+    "showcase.sortLabel": "Ordina la vetrina",
+    "showcase.sort.newest": "Prima i più recenti",
+    "showcase.sort.oldest": "Prima i meno recenti",
+    "showcase.sort.nameAsc": "Nome (A–Z)",
+    "showcase.sort.nameDesc": "Nome (Z–A)",
+    "showcase.empty": "Nessuna app corrisponde alla tua ricerca.",
+    "showcase.paginationLabel": "Paginazione della vetrina",
+    "showcase.cta.title": "Usi Vapor nella tua app? Ci piacerebbe metterla in evidenza!",
+    "showcase.cta.caption": "Aiutaci a mostrare i tanti modi diversi in cui si può usare Vapor.",
+    "showcase.cta.button": "Mettiti in contatto",
+    "footer.joinDiscord": "Unisciti al nostro Discord",
+    // Nav / footer chrome.
+    "nav.documentation": "Documentazione",
+    "nav.frameworkDocs": "Documentazione del Framework",
+    "nav.frameworkDocs.caption": "Impara a usare Vapor",
+    "nav.apiDocs": "Documentazione API",
+    "nav.apiDocs.caption": "Documentazione di riferimento per Vapor",
+    "nav.selectTheme": "Seleziona il tema",
+    "nav.closeMenu": "Chiudi il menu",
+    "footer.supporters": "Sostenitori",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} stelle su GitHub",
+    "home.hero.starsLoading": "… stelle su GitHub",
+    "home.companies.caption": "Al servizio di aziende come:",
+    "home.packages.title": "Sviluppa lato server con gli strumenti che già conosci",
+    "home.packages.caption": "Non c’è bisogno di imparare un linguaggio da zero (o di mettere insieme un altro team) solo per il tuo backend. Vapor è costruito su SwiftNIO di Apple, così potrai usare il linguaggio che già conosci e ami.",
+    "home.packages.vapor.description": "Crea API efficienti in un linguaggio che ami. Definisci rotte, invia e ricevi JSON e costruisci server HTTP.",
+    "home.packages.fluent.description": "Crea modelli e interagisci con il tuo database in codice Swift nativo e sicuro, senza dover scrivere alcun SQL.",
+    "home.packages.jwt.description": "Crea, firma e verifica i JSON Web Token in Swift. Costruito su Swift Crypto.",
+    "home.packages.leaf.description": "Un motore di template scritto in Swift. Genera HTML per app web ed email con una sintassi semplice che chiunque può usare.",
+    "home.features.performance.title": "API e server ad alte prestazioni",
+    "home.features.performance.description": "Costruito con un’architettura non bloccante e orientata agli eventi, Vapor ti permette di creare API e server HTTP scalabili e ad alte prestazioni. Grazie al modello di concorrenza di Swift, puoi scrivere codice chiaro e manutenibile, efficiente e facile da leggere.",
+    "home.features.confidence.title": "Rilascia con sicurezza, anche di venerdì",
+    "home.features.confidence.description": "Grazie al design espressivo e orientato ai protocolli di Vapor, potrai rilasciare il tuo codice con serenità. Con la nostra forte attenzione alla type-safety, molti errori e problemi vengono individuati subito dal compilatore.",
+    "home.integrations.label": "Integrazioni",
+    "home.integrations.title": "Integrazione completa con il tuo intero stack",
+    "home.integrations.description": "Il maturo ecosistema di Vapor include oltre cento pacchetti Swift orientati al server, ufficiali e mantenuti dalla community, per rendere lo sviluppo delle tue applicazioni semplice ed efficiente.",
+    "home.showcase.title": "Usato dai team dietro queste fantastiche app",
+    "home.sponsors.title": "Sostenuto dai nostri incredibili sponsor e sostenitori",
+    "home.discord.title": "Unisciti alla più grande community di sviluppatori backend Swift",
+    "home.discord.description": "La community Discord di Vapor, con oltre 13.000 membri, sarà al tuo fianco per supportarti lungo il percorso. Fai domande, contribuisci e diventa parte di un angolo di internet vivace e accogliente.",
+    "home.apiDocs.title": "Ottieni tutti gli strumenti di cui hai bisogno per sviluppare con Swift",
+    "home.apiDocs.item1": "Crea facilmente nuovi progetti con la Vapor Toolbox",
+    "home.apiDocs.item2": "Documentazione completa e riferimento delle API",
+    "home.apiDocs.item3": "Tutto ciò di cui hai bisogno per creare backend e API",
+    "home.apiDocs.item4": "Supporto completo per il modello di concorrenza di Swift",
+    "comingSoon.message": "Questa pagina è ancora in costruzione, resta sintonizzato per gli aggiornamenti!",
+    "pages.showcase.title": "Vetrina",
+    "pages.team.title": "Team",
+    "pages.supporters.title": "Sostenitori",
+    "pages.evangelists.title": "Evangelisti",
+]
+
+let italianLanguage = Language(
+    .italian,
+    customStrings: italianStrings,
+    image: "static/images/opengraph/vapor-og-it-2x.png",
+    localisation: .init(
+        notFoundTitle: "Pagina non trovata",
+        notFoundMessage: "La pagina che stai cercando potrebbe essere stata spostata, rinominata o non essere mai esistita.",
+        notFoundLink: "Torna alla home page",
+        toggleNavigation: "Attiva/disattiva la navigazione",
+        skipToContent: "Vai al contenuto",
+        home: "Home",
+        store: "Shop",
+        blog: "Blog",
+        showcase: "Vetrina",
+        team: "Team",
+        help: "Aiuto",
+        pressKit: "Kit stampa",
+        community: "Community",
+        resources: "Risorse",
+        language: "Lingua",
+        theme: "Tema",
+        lightTheme: "Chiaro",
+        darkTheme: "Scuro",
+        systemTheme: "Sistema"
+    )
+)
