+++ Sources/VaporWebsite/Translations/Japanese.swift
@@ -1,0 +1,102 @@
+// Japanese strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let japaneseStrings: [String: String] = [
+    "home.hero.title": "Swift、ただしサーバー上で",
+    "home.hero.caption": "Vaporは、SwiftでHTTPサーバー、バックエンド、APIを構築するための安全で高性能、かつ使いやすい基盤を提供します",
+    // Buttons / controls.
+    "cta.getStarted": "始める",
+    "cta.learnMore": "詳しく見る",
+    "cta.seeInAction": "実際の動作を見る",
+    "home.apiDocs.goToApiDocs": "APIドキュメントへ",
+    "home.discord.contribute": "貢献する",
+    "home.showcase.seeFullList": "一覧を見る",
+    "home.showcase.prev": "前のショーケース",
+    "home.showcase.next": "次のショーケース",
+    "home.sponsors.becomeSupporter": "サポーターになる",
+    "home.sponsors.prev": "前のスポンサー",
+    "home.sponsors.next": "次のスポンサー",
+    "showcase.title": "Vapor は高性能な数千ものアプリやサイトを支えています",
+    "showcase.caption": "Vapor で構築されたアプリ、サイト、企業、そしてそれを支えるチームをご覧ください。",
+    "showcase.search": "検索",
+    "showcase.searchLabel": "ショーケースを検索",
+    "showcase.sortLabel": "ショーケースを並べ替え",
+    "showcase.sort.newest": "新しい順",
+    "showcase.sort.oldest": "古い順",
+    "showcase.sort.nameAsc": "名前（A–Z）",
+    "showcase.sort.nameDesc": "名前（Z–A）",
+    "showcase.empty": "検索に一致するアプリはありません。",
+    "showcase.paginationLabel": "ショーケースのページ送り",
+    "showcase.cta.title": "あなたのアプリで Vapor を使っていますか？ぜひご紹介させてください！",
+    "showcase.cta.caption": "Vapor のさまざまな活用方法を一緒に紹介しましょう。",
+    "showcase.cta.button": "お問い合わせ",
+    "footer.joinDiscord": "Discordに参加する",
+    // Nav / footer chrome.
+    "nav.documentation": "ドキュメント",
+    "nav.frameworkDocs": "フレームワークドキュメント",
+    "nav.frameworkDocs.caption": "Vaporの使い方を学ぶ",
+    "nav.apiDocs": "APIドキュメント",
+    "nav.apiDocs.caption": "Vaporのリファレンスドキュメント",
+    "nav.selectTheme": "テーマを選択",
+    "nav.closeMenu": "メニューを閉じる",
+    "footer.supporters": "サポーター",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "GitHubで{count}個のスター",
+    "home.hero.starsLoading": "GitHubで…個のスター",
+    "home.companies.caption": "次のような企業で利用されています:",
+    "home.packages.title": "すでに慣れ親しんだツールでサーバーサイドを構築",
+    "home.packages.caption": "バックエンドのためだけに、言語をゼロから学んだり（別のチームを編成したり）する必要はありません。VaporはAppleのSwiftNIO上に構築されているため、すでに知っていて愛用している言語をそのまま使えます。",
+    "home.packages.vapor.description": "好きな言語で効率的なAPIを構築。ルートを作成し、JSONを送受信し、HTTPサーバーを構築できます。",
+    "home.packages.fluent.description": "ネイティブで安全なSwiftコードでモデルを作成し、SQLを書くことなくデータベースを操作できます。",
+    "home.packages.jwt.description": "SwiftでJSON Web Tokenを作成、署名、検証。Swift Cryptoを基盤としています。",
+    "home.packages.leaf.description": "Swiftで書かれたテンプレートエンジン。誰でも使えるシンプルな構文で、WebアプリとメールのHTMLを生成できます。",
+    "home.features.performance.title": "高性能なAPIとサーバー",
+    "home.features.performance.description": "ノンブロッキングかつイベント駆動のアーキテクチャで構築されたVaporなら、高性能でスケーラブルなAPIとHTTPサーバーを構築できます。SwiftのConcurrencyモデルを使えば、効率的で読みやすく、明確で保守しやすいコードを書けます。",
+    "home.features.confidence.title": "金曜日でも自信を持ってリリース",
+    "home.features.confidence.description": "Vaporの表現力豊かでプロトコル指向の設計により、安心してコードをリリースできます。型安全性を重視しているため、多くのエラーや問題はコンパイラによって早い段階で検出されます。",
+    "home.integrations.label": "連携",
+    "home.integrations.title": "スタック全体とのフル連携",
+    "home.integrations.description": "Vaporの成熟したエコシステムには、公式およびコミュニティが管理する100を超えるサーバー向けSwiftパッケージが含まれており、アプリケーション開発を簡単かつ効率的にします。",
+    "home.showcase.title": "これらの素晴らしいアプリを手がけるチームが利用しています",
+    "home.sponsors.title": "素晴らしいスポンサーと支援者の皆様に支えられています",
+    "home.discord.title": "Swiftバックエンド開発者の最大のコミュニティに参加しよう",
+    "home.discord.description": "13,000人を超えるVaporのDiscordコミュニティが、あなたの歩みをサポートします。質問したり、貢献したりして、活気にあふれた居心地のよいインターネットの一角の一員になりましょう。",
+    "home.apiDocs.title": "Swiftでの開発に必要なすべてのツールを手に入れよう",
+    "home.apiDocs.item1": "Vapor Toolboxで新しいプロジェクトを簡単に作成",
+    "home.apiDocs.item2": "充実したドキュメントとAPIリファレンス",
+    "home.apiDocs.item3": "バックエンドとAPIの構築に必要なすべて",
+    "home.apiDocs.item4": "SwiftのConcurrencyモデルを完全サポート",
+    "comingSoon.message": "このページはまだ準備中です。最新情報をお待ちください！",
+    "pages.showcase.title": "ショーケース",
+    "pages.team.title": "チーム",
+    "pages.supporters.title": "サポーター",
+    "pages.evangelists.title": "エヴァンジェリスト",
+]
+
+let japaneseLanguage = Language(
+    .japanese,
+    customStrings: japaneseStrings,
+    image: "static/images/opengraph/vapor-og-ja-2x.png",
+    localisation: .init(
+        notFoundTitle: "ページが見つかりません",
+        notFoundMessage: "お探しのページは移動または名称変更されたか、存在しない可能性があります。",
+        notFoundLink: "ホームページに戻る",
+        toggleNavigation: "ナビゲーションの切り替え",
+        skipToContent: "コンテンツにスキップ",
+        home: "ホーム",
+        store: "ストア",
+        blog: "ブログ",
+        showcase: "ショーケース",
+        team: "チーム",
+        help: "ヘルプ",
+        pressKit: "プレスキット",
+        community: "コミュニティ",
+        resources: "リソース",
+        language: "言語",
+        theme: "テーマ",
+        lightTheme: "ライト",
+        darkTheme: "ダーク",
+        systemTheme: "システム"
+    )
+)
