+++ Sources/VaporWebsite/Translations/Korean.swift
@@ -1,0 +1,102 @@
+// Korean strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let koreanStrings: [String: String] = [
+    "home.hero.title": "Swift, 하지만 서버에서",
+    "home.hero.caption": "Vapor는 Swift로 HTTP 서버, 백엔드 및 API를 구축할 수 있는 안전하고 성능이 뛰어나며 사용하기 쉬운 기반을 제공합니다",
+    // Buttons / controls.
+    "cta.getStarted": "시작하기",
+    "cta.learnMore": "자세히 보기",
+    "cta.seeInAction": "직접 확인하기",
+    "home.apiDocs.goToApiDocs": "API 문서로 이동",
+    "home.discord.contribute": "기여하기",
+    "home.showcase.seeFullList": "전체 목록 보기",
+    "home.showcase.prev": "이전 쇼케이스",
+    "home.showcase.next": "다음 쇼케이스",
+    "home.sponsors.becomeSupporter": "후원자 되기",
+    "home.sponsors.prev": "이전 후원자",
+    "home.sponsors.next": "다음 후원자",
+    "showcase.title": "Vapor는 수천 개의 고성능 앱과 사이트를 지원합니다",
+    "showcase.caption": "Vapor로 만들어진 앱, 사이트, 기업과 그 뒤의 팀들을 만나보세요.",
+    "showcase.search": "검색",
+    "showcase.searchLabel": "쇼케이스 검색",
+    "showcase.sortLabel": "쇼케이스 정렬",
+    "showcase.sort.newest": "최신순",
+    "showcase.sort.oldest": "오래된순",
+    "showcase.sort.nameAsc": "이름 (A–Z)",
+    "showcase.sort.nameDesc": "이름 (Z–A)",
+    "showcase.empty": "검색과 일치하는 앱이 없습니다.",
+    "showcase.paginationLabel": "쇼케이스 페이지 매김",
+    "showcase.cta.title": "앱에서 Vapor를 사용하고 계신가요? 소개해 드리고 싶어요!",
+    "showcase.cta.caption": "Vapor를 활용하는 다양한 방법을 함께 보여주세요.",
+    "showcase.cta.button": "문의하기",
+    "footer.joinDiscord": "Discord에 참여하기",
+    // Nav / footer chrome.
+    "nav.documentation": "문서",
+    "nav.frameworkDocs": "프레임워크 문서",
+    "nav.frameworkDocs.caption": "Vapor 사용법 알아보기",
+    "nav.apiDocs": "API 문서",
+    "nav.apiDocs.caption": "Vapor 레퍼런스 문서",
+    "nav.selectTheme": "테마 선택",
+    "nav.closeMenu": "메뉴 닫기",
+    "footer.supporters": "후원자",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "GitHub에서 {count}개의 스타",
+    "home.hero.starsLoading": "GitHub에서 …개의 스타",
+    "home.companies.caption": "다음과 같은 기업들이 사용합니다:",
+    "home.packages.title": "이미 알고 있는 도구로 서버 사이드를 구축하세요",
+    "home.packages.caption": "백엔드만을 위해 언어를 처음부터 배우거나 별도의 팀을 꾸릴 필요가 없습니다. Vapor는 Apple의 SwiftNIO를 기반으로 만들어져, 이미 알고 좋아하는 언어를 그대로 사용할 수 있습니다.",
+    "home.packages.vapor.description": "좋아하는 언어로 효율적인 API를 구축하세요. 라우트를 만들고 JSON을 주고받으며 HTTP 서버를 구축할 수 있습니다.",
+    "home.packages.fluent.description": "네이티브하고 안전한 Swift 코드로 모델을 만들고, SQL을 작성할 필요 없이 데이터베이스와 상호작용하세요.",
+    "home.packages.jwt.description": "Swift에서 JSON Web Token을 생성, 서명, 검증하세요. Swift Crypto를 기반으로 합니다.",
+    "home.packages.leaf.description": "Swift로 작성된 템플릿 엔진. 누구나 사용할 수 있는 간단한 문법으로 웹 앱과 이메일용 HTML을 생성하세요.",
+    "home.features.performance.title": "고성능 API와 서버",
+    "home.features.performance.description": "논블로킹 이벤트 기반 아키텍처로 구축된 Vapor를 사용하면 고성능의 확장 가능한 API와 HTTP 서버를 만들 수 있습니다. Swift의 동시성 모델을 활용해 효율적이고 읽기 쉬우며 명확하고 유지보수하기 좋은 코드를 작성할 수 있습니다.",
+    "home.features.confidence.title": "금요일에도 자신 있게 배포하세요",
+    "home.features.confidence.description": "Vapor의 표현력 있고 프로토콜 지향적인 설계 덕분에 안심하고 코드를 배포할 수 있습니다. 타입 안전성에 중점을 두어 많은 오류와 문제를 컴파일러가 일찍 잡아냅니다.",
+    "home.integrations.label": "통합",
+    "home.integrations.title": "전체 스택과의 완벽한 통합",
+    "home.integrations.description": "Vapor의 성숙한 생태계에는 공식 및 커뮤니티가 관리하는 100개가 넘는 서버 우선 Swift 패키지가 포함되어 있어, 애플리케이션 개발을 쉽고 효율적으로 만들어 줍니다.",
+    "home.showcase.title": "이 멋진 앱들을 만든 팀들이 사용합니다",
+    "home.sponsors.title": "멋진 후원자와 서포터 여러분의 지원을 받고 있습니다",
+    "home.discord.title": "최대 규모의 Swift 백엔드 개발자 커뮤니티에 참여하세요",
+    "home.discord.description": "13,000명이 넘는 Vapor의 Discord 커뮤니티가 여러분의 여정을 함께합니다. 질문하고, 기여하며, 활기차고 따뜻한 인터넷 공간의 일원이 되어 보세요.",
+    "home.apiDocs.title": "Swift로 개발하는 데 필요한 모든 도구를 만나보세요",
+    "home.apiDocs.item1": "Vapor Toolbox로 손쉽게 새 프로젝트를 만드세요",
+    "home.apiDocs.item2": "방대한 문서와 API 레퍼런스",
+    "home.apiDocs.item3": "백엔드와 API 구축에 필요한 모든 것",
+    "home.apiDocs.item4": "Swift의 동시성 모델 완벽 지원",
+    "comingSoon.message": "이 페이지는 아직 준비 중입니다. 업데이트를 기대해 주세요!",
+    "pages.showcase.title": "쇼케이스",
+    "pages.team.title": "팀",
+    "pages.supporters.title": "후원자",
+    "pages.evangelists.title": "에반젤리스트",
+]
+
+let koreanLanguage = Language(
+    .korean,
+    customStrings: koreanStrings,
+    image: "static/images/opengraph/vapor-og-ko-2x.png",
+    localisation: .init(
+        notFoundTitle: "페이지를 찾을 수 없습니다",
+        notFoundMessage: "찾고 있는 페이지가 이동되었거나 이름이 변경되었거나 존재하지 않을 수 있습니다.",
+        notFoundLink: "홈페이지로 돌아가기",
+        toggleNavigation: "내비게이션 전환",
+        skipToContent: "본문으로 건너뛰기",
+        home: "홈",
+        store: "스토어",
+        blog: "블로그",
+        showcase: "쇼케이스",
+        team: "팀",
+        help: "도움말",
+        pressKit: "프레스 킷",
+        community: "커뮤니티",
+        resources: "리소스",
+        language: "언어",
+        theme: "테마",
+        lightTheme: "라이트",
+        darkTheme: "다크",
+        systemTheme: "시스템"
+    )
+)
