+++ Sources/VaporWebsite/Translations/Polish.swift
@@ -1,0 +1,102 @@
+// Polish strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let polishStrings: [String: String] = [
+    "home.hero.title": "Swift, ale na serwerze",
+    "home.hero.caption": "Vapor zapewnia bezpieczną, wydajną i łatwą w użyciu podstawę do tworzenia serwerów HTTP, backendów i interfejsów API w Swift",
+    // Buttons / controls.
+    "cta.getStarted": "Rozpocznij",
+    "cta.learnMore": "Dowiedz się więcej",
+    "cta.seeInAction": "Zobacz w działaniu",
+    "home.apiDocs.goToApiDocs": "Przejdź do dokumentacji API",
+    "home.discord.contribute": "Współtwórz",
+    "home.showcase.seeFullList": "Zobacz pełną listę",
+    "home.showcase.prev": "Poprzedni showcase",
+    "home.showcase.next": "Następny showcase",
+    "home.sponsors.becomeSupporter": "Zostań sponsorem",
+    "home.sponsors.prev": "Poprzedni sponsorzy",
+    "home.sponsors.next": "Następni sponsorzy",
+    "showcase.title": "Vapor napędza tysiące wydajnych aplikacji i witryn",
+    "showcase.caption": "Poznaj aplikacje, witryny i firmy oparte na Vaporze – oraz stojące za nimi zespoły.",
+    "showcase.search": "Szukaj",
+    "showcase.searchLabel": "Przeszukaj realizacje",
+    "showcase.sortLabel": "Sortuj realizacje",
+    "showcase.sort.newest": "Najnowsze najpierw",
+    "showcase.sort.oldest": "Najstarsze najpierw",
+    "showcase.sort.nameAsc": "Nazwa (A–Z)",
+    "showcase.sort.nameDesc": "Nazwa (Z–A)",
+    "showcase.empty": "Żadna aplikacja nie pasuje do wyszukiwania.",
+    "showcase.paginationLabel": "Paginacja realizacji",
+    "showcase.cta.title": "Używasz Vapora w swojej aplikacji? Chętnie ją wyróżnimy!",
+    "showcase.cta.caption": "Pomóż pokazać, na ile różnych sposobów można używać Vapora.",
+    "showcase.cta.button": "Skontaktuj się",
+    "footer.joinDiscord": "Dołącz do naszego Discorda",
+    // Nav / footer chrome.
+    "nav.documentation": "Dokumentacja",
+    "nav.frameworkDocs": "Dokumentacja frameworka",
+    "nav.frameworkDocs.caption": "Naucz się korzystać z Vapor",
+    "nav.apiDocs": "Dokumentacja API",
+    "nav.apiDocs.caption": "Dokumentacja referencyjna Vapor",
+    "nav.selectTheme": "Wybierz motyw",
+    "nav.closeMenu": "Zamknij menu",
+    "footer.supporters": "Wspierający",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} gwiazdek na GitHub",
+    "home.hero.starsLoading": "… gwiazdek na GitHub",
+    "home.companies.caption": "Napędza firmy takie jak:",
+    "home.packages.title": "Twórz po stronie serwera za pomocą narzędzi, które już znasz",
+    "home.packages.caption": "Nie musisz uczyć się języka od zera (ani budować osobnego zespołu) tylko dla swojego backendu. Vapor jest zbudowany na bazie SwiftNIO firmy Apple, więc możesz używać języka, który już znasz i lubisz.",
+    "home.packages.vapor.description": "Twórz wydajne API w języku, który kochasz. Definiuj trasy, wysyłaj i odbieraj JSON oraz buduj serwery HTTP.",
+    "home.packages.fluent.description": "Twórz modele i korzystaj z bazy danych w natywnym, bezpiecznym kodzie Swift, bez konieczności pisania SQL.",
+    "home.packages.jwt.description": "Twórz, podpisuj i weryfikuj tokeny JSON Web Token w Swift. Zbudowane na Swift Crypto.",
+    "home.packages.leaf.description": "Silnik szablonów napisany w Swift. Generuj HTML dla aplikacji webowych i e-maili dzięki prostej składni, z której skorzysta każdy.",
+    "home.features.performance.title": "Wydajne API i serwery",
+    "home.features.performance.description": "Zbudowany w oparciu o nieblokującą architekturę sterowaną zdarzeniami, Vapor pozwala tworzyć wydajne i skalowalne API oraz serwery HTTP. Dzięki modelowi współbieżności Swift możesz pisać przejrzysty, łatwy w utrzymaniu kod, który jest wydajny i czytelny.",
+    "home.features.confidence.title": "Wdrażaj z pewnością, nawet w piątki",
+    "home.features.confidence.description": "Dzięki ekspresyjnemu, zorientowanemu na protokoły projektowi Vapora wdrażasz swój kod ze spokojem. Dzięki naszemu silnemu naciskowi na bezpieczeństwo typów wiele błędów i problemów jest wychwytywanych wcześnie przez kompilator.",
+    "home.integrations.label": "Integracje",
+    "home.integrations.title": "Pełna integracja z całym Twoim stosem",
+    "home.integrations.description": "Dojrzały ekosystem Vapora obejmuje ponad sto oficjalnych i utrzymywanych przez społeczność pakietów Swift zorientowanych na serwer, dzięki którym tworzenie aplikacji jest łatwe i wydajne.",
+    "home.showcase.title": "Używany przez zespoły stojące za tymi wspaniałymi aplikacjami",
+    "home.sponsors.title": "Wspierany przez naszych niesamowitych sponsorów i darczyńców",
+    "home.discord.title": "Dołącz do największej społeczności programistów backendu w Swift",
+    "home.discord.description": "Licząca ponad 13 000 członków społeczność Discord Vapora będzie u Twojego boku, by wspierać Cię na każdym kroku. Zadawaj pytania, współtwórz i stań się częścią tętniącego życiem, przyjaznego zakątka internetu.",
+    "home.apiDocs.title": "Zdobądź wszystkie narzędzia potrzebne do tworzenia w Swift",
+    "home.apiDocs.item1": "Łatwo twórz nowe projekty dzięki Vapor Toolbox",
+    "home.apiDocs.item2": "Obszerna dokumentacja i dokumentacja API",
+    "home.apiDocs.item3": "Wszystko, czego potrzebujesz do tworzenia backendów i API",
+    "home.apiDocs.item4": "Pełne wsparcie dla modelu współbieżności Swift",
+    "comingSoon.message": "Ta strona jest jeszcze w budowie, śledź aktualizacje!",
+    "pages.showcase.title": "Realizacje",
+    "pages.team.title": "Zespół",
+    "pages.supporters.title": "Wspierający",
+    "pages.evangelists.title": "Ewangeliści",
+]
+
+let polishLanguage = Language(
+    .polish,
+    customStrings: polishStrings,
+    image: "static/images/opengraph/vapor-og-pl-2x.png",
+    localisation: .init(
+        notFoundTitle: "Nie znaleziono strony",
+        notFoundMessage: "Strona, której szukasz, mogła zostać przeniesiona, zmieniona lub nigdy nie istniała.",
+        notFoundLink: "Powrót do strony głównej",
+        toggleNavigation: "Przełącz nawigację",
+        skipToContent: "Przejdź do treści",
+        home: "Strona główna",
+        store: "Sklep",
+        blog: "Blog",
+        showcase: "Realizacje",
+        team: "Zespół",
+        help: "Pomoc",
+        pressKit: "Materiały prasowe",
+        community: "Społeczność",
+        resources: "Zasoby",
+        language: "Język",
+        theme: "Motyw",
+        lightTheme: "Jasny",
+        darkTheme: "Ciemny",
+        systemTheme: "Systemowy"
+    )
+)
