+++ Sources/VaporWebsite/Translations/Spanish.swift
@@ -1,0 +1,102 @@
+// Spanish strings + Language for the Vapor website. Copy the keys from
+// englishStrings (in English.swift) — anything omitted falls back to English.
+import Kiln
+
+let spanishStrings: [String: String] = [
+    "home.hero.title": "Swift, pero en el servidor",
+    "home.hero.caption": "Vapor proporciona una base segura, eficiente y fácil de usar para crear servidores HTTP, backends y APIs en Swift",
+    // Buttons / controls.
+    "cta.getStarted": "Comenzar",
+    "cta.learnMore": "Más información",
+    "cta.seeInAction": "Míralo en acción",
+    "home.apiDocs.goToApiDocs": "Ir a la documentación de la API",
+    "home.discord.contribute": "Contribuir",
+    "home.showcase.seeFullList": "Ver lista completa",
+    "home.showcase.prev": "Showcase anterior",
+    "home.showcase.next": "Showcase siguiente",
+    "home.sponsors.becomeSupporter": "Conviértete en patrocinador",
+    "home.sponsors.prev": "Patrocinadores anteriores",
+    "home.sponsors.next": "Patrocinadores siguientes",
+    "showcase.title": "Vapor impulsa miles de apps y sitios de alto rendimiento",
+    "showcase.caption": "Descubre las apps, los sitios y las empresas que construyen sobre Vapor, y los equipos que están detrás.",
+    "showcase.search": "Buscar",
+    "showcase.searchLabel": "Buscar en los proyectos",
+    "showcase.sortLabel": "Ordenar los proyectos",
+    "showcase.sort.newest": "Más recientes primero",
+    "showcase.sort.oldest": "Más antiguos primero",
+    "showcase.sort.nameAsc": "Nombre (A–Z)",
+    "showcase.sort.nameDesc": "Nombre (Z–A)",
+    "showcase.empty": "Ningún proyecto coincide con tu búsqueda.",
+    "showcase.paginationLabel": "Paginación de proyectos",
+    "showcase.cta.title": "¿Usas Vapor en tu app? ¡Nos encantaría destacarla!",
+    "showcase.cta.caption": "Ayuda a demostrar las distintas formas en que se puede usar Vapor.",
+    "showcase.cta.button": "Ponte en contacto",
+    "footer.joinDiscord": "Únete a nuestro Discord",
+    // Nav / footer chrome.
+    "nav.documentation": "Documentación",
+    "nav.frameworkDocs": "Documentación del framework",
+    "nav.frameworkDocs.caption": "Aprende a usar Vapor",
+    "nav.apiDocs": "Documentación de la API",
+    "nav.apiDocs.caption": "Documentación de referencia de Vapor",
+    "nav.selectTheme": "Seleccionar tema",
+    "nav.closeMenu": "Cerrar el menú",
+    "footer.supporters": "Colaboradores",
+    // Home — stars / companies / packages.
+    "home.hero.stars": "{count} estrellas en GitHub",
+    "home.hero.starsLoading": "… estrellas en GitHub",
+    "home.companies.caption": "Impulsando a empresas como:",
+    "home.packages.title": "Desarrolla el lado del servidor con las herramientas que ya conoces",
+    "home.packages.caption": "No necesitas aprender un lenguaje desde cero (ni formar un equipo distinto) solo para tu backend. Vapor está construido sobre SwiftNIO de Apple, así que podrás usar el lenguaje que ya conoces y adoras.",
+    "home.packages.vapor.description": "Crea APIs eficientes en un lenguaje que te encanta. Define rutas, envía y recibe JSON y construye servidores HTTP.",
+    "home.packages.fluent.description": "Crea modelos e interactúa con tu base de datos en código Swift nativo y seguro, sin necesidad de escribir SQL.",
+    "home.packages.jwt.description": "Crea, firma y verifica JSON Web Tokens en Swift. Construido sobre Swift Crypto.",
+    "home.packages.leaf.description": "Un motor de plantillas escrito en Swift. Genera HTML para aplicaciones web y correos electrónicos con una sintaxis sencilla que cualquiera puede usar.",
+    "home.features.performance.title": "APIs y servidores de alto rendimiento",
+    "home.features.performance.description": "Construido con una arquitectura no bloqueante y orientada a eventos, Vapor te permite crear APIs y servidores HTTP escalables y de alto rendimiento. Gracias al modelo de concurrencia de Swift, puedes escribir código claro y mantenible que es eficiente y fácil de leer.",
+    "home.features.confidence.title": "Lanza con confianza, incluso los viernes",
+    "home.features.confidence.description": "Con el diseño expresivo y orientado a protocolos de Vapor, tendrás tranquilidad al lanzar tu código. Gracias a nuestro enfoque en la seguridad de tipos, el compilador detecta muchos errores y problemas desde el principio.",
+    "home.integrations.label": "Integraciones",
+    "home.integrations.title": "Integración total con todo tu stack",
+    "home.integrations.description": "El maduro ecosistema de Vapor incluye más de cien paquetes Swift orientados al servidor, oficiales y mantenidos por la comunidad, para que crear tus aplicaciones sea fácil y eficiente.",
+    "home.showcase.title": "Usado por los equipos detrás de estas increíbles apps",
+    "home.sponsors.title": "Con el apoyo de nuestros increíbles patrocinadores y colaboradores",
+    "home.discord.title": "Únete a la mayor comunidad de desarrolladores backend de Swift",
+    "home.discord.description": "La comunidad de Discord de Vapor, con más de 13.000 miembros, estará a tu lado para ayudarte en el camino. Haz preguntas, contribuye y forma parte de un rincón próspero y acogedor de internet.",
+    "home.apiDocs.title": "Obtén todas las herramientas que necesitas para desarrollar con Swift",
+    "home.apiDocs.item1": "Crea nuevos proyectos fácilmente con Vapor Toolbox",
+    "home.apiDocs.item2": "Amplia documentación y referencia de la API",
+    "home.apiDocs.item3": "Todo lo que necesitas para crear backends y APIs",
+    "home.apiDocs.item4": "Compatibilidad total con el modelo de concurrencia de Swift",
+    "comingSoon.message": "Esta página aún está en construcción. ¡Vuelve pronto para ver novedades!",
+    "pages.showcase.title": "Proyectos",
+    "pages.team.title": "Equipo",
+    "pages.supporters.title": "Colaboradores",
+    "pages.evangelists.title": "Evangelistas",
+]
+
+let spanishLanguage = Language(
+    .spanish,
+    customStrings: spanishStrings,
+    image: "static/images/opengraph/vapor-og-es-2x.png",
+    localisation: .init(
+        notFoundTitle: "Página no encontrada",
+        notFoundMessage: "Es posible que la página que buscas se haya movido, cambiado de nombre o que nunca haya existido.",
+        notFoundLink: "Volver a la página de inicio",
+        toggleNavigation: "Alternar la navegación",
+        skipToContent: "Saltar al contenido",
+        home: "Inicio",
+        store: "Tienda",
+        blog: "Blog",
+        showcase: "Proyectos",
+        team: "Equipo",
+        help: "Ayuda",
+        pressKit: "Kit de prensa",
+        community: "Comunidad",
+        resources: "Recursos",
+        language: "Idioma",
+        theme: "Tema",
+        lightTheme: "Claro",
+        darkTheme: "Oscuro",
+        systemTheme: "Sistema"
+    )
+)
