+++ Sources/VaporWebsite/main.swift
@@ -2,825 +2,14 @@ import Kiln
 
 // The Vapor marketing site (vapor.codes), built with Kiln.
 //
-// Unlike the docs site, these pages are custom-HTML rather than prose, so the
-// markup lives in Leaf templates under `Theme/` (a custom theme that fully
-// overrides the bundled one). Each content file is a thin front-matter stub that
-// selects a template; the home page (`index.md`, isHome) uses `home.leaf`.
+// The pages are custom HTML in Leaf templates under `Theme/` (a custom theme
+// overriding the bundled one); each `Content/*.md` is a thin front-matter stub
+// selecting a template (the home page uses `home.leaf`).
 //
-// Localisation: all visible copy is either a built-in chrome string (`strings.*`)
-// or a theme-defined `#localise("key")` looked up in `customStrings`. English is
-// the default language, so its `customStrings` below double as the fallback for
-// any key a translation omits. No per-language `.md` files are needed — the
-// `/de/` pages are built from the same templates with the German strings.
-
-// MARK: - Theme strings (English = default + fallback)
-
-// Every `#localise("…")` key used in the templates must appear here, or the raw
-// key would render. These are the exact English strings, so the English build is
-// byte-for-byte the old Publish output.
-let englishStrings: [String: String] = [
-    // Navigation / footer chrome not covered by the built-in `strings.*` set.
-    "nav.closeMenu": "Close menu",
-    "nav.documentation": "Documentation",
-    "nav.frameworkDocs": "Framework Docs",
-    "nav.frameworkDocs.caption": "Learn how to use Vapor",
-    "nav.apiDocs": "API Docs",
-    "nav.apiDocs.caption": "Reference documentation for Vapor",
-    "nav.selectTheme": "Select theme",
-    "footer.joinDiscord": "Join our Discord",
-    "footer.supporters": "Supporters",
-    "comingSoon.message": "This page is still under construction, stay tuned for updates!",
-    // Stub ("coming soon") page titles — drive the localised H1 + <title> via each
-    // page's `titleKey` front-matter field. (Some of these pages will get real
-    // implementations later; these keep the placeholders localised in the meantime.)
-    "pages.showcase.title": "Showcase",
-    "pages.team.title": "Team",
-    "pages.supporters.title": "Supporters",
-    "pages.evangelists.title": "Evangelists",
-    // Calls to action (shared across sections).
-    "cta.getStarted": "Get Started",
-    "cta.learnMore": "Learn More",
-    "cta.seeInAction": "See it in action",
-    // Home — hero.
-    "home.hero.title": "Swift, but on a server",
-    "home.hero.caption": "Vapor provides a safe, performant and easy to use foundation to build HTTP servers, backends and APIs in Swift",
-    // `{count}` is filled in by updateStarsCount.js with the live GitHub star count.
-    "home.hero.stars": "{count} stars on GitHub",
-    "home.hero.starsLoading": "… stars on GitHub",
-    // Home — companies.
-    "home.companies.caption": "Powering companies like:",
-    // Home — packages.
-    "home.packages.title": "Build server side with the tooling you already understand",
-    "home.packages.caption": "No need to learn a language from scratch (or assemble a different team) just for your backend. Vapor is built on Apple’s SwiftNIO so you’ll get to use the language you already know and love.",
-    // Package feature cards (the card titles — Vapor/Fluent/JWT/Leaf — are product
-    // names and stay untranslated; only the descriptions are localised).
-    "home.packages.vapor.description": "Build efficient APIs in a language you love. Create routes, send and receive JSON and build HTTP servers.",
-    "home.packages.fluent.description": "Create models and interact with your database in native, safe Swift code without needing to write any SQL",
-    "home.packages.jwt.description": "Create, sign and verify JSON Web Tokens in Swift. Built on top of Swift Crypto",
-    "home.packages.leaf.description": "A templating engine written in Swift. Generate HTML for both web apps and emails with a simple syntax anyone can use",
-    // Home — features.
-    "home.features.performance.title": "High-performant APIs and servers",
-    "home.features.performance.description": "Built with a non-blocking, event-driven architecture, Vapor allows you to build high-performant, scalable APIs and HTTP servers. Using Swift's Concurrency model, you can write clear, maintainable code that's efficient and easy to read.",
-    "home.features.confidence.title": "Ship with confidence, even on Fridays",
-    "home.features.confidence.description": "With Vapor's expressive, protocol oriented design, you'll have peace of mind when shipping your code. With our strong type-safety focus, many errors and problems are caught early on by the compiler.",
-    // Home — integrations.
-    "home.integrations.label": "Integrations",
-    "home.integrations.title": "Full integration with your entire stack",
-    "home.integrations.description": "Vapor's mature ecosystem includes over a hundred official and community maintained server-first Swift packages to make building your applications easy and efficient.",
-    // Home — showcase.
-    "home.showcase.title": "Used by the teams behind these amazing apps",
-    "home.showcase.seeFullList": "See full list",
-    "home.showcase.prev": "Previous Showcase",
-    "home.showcase.next": "Next Showcase",
-    // Home — discord.
-    "home.discord.title": "Join the largest community of Swift backend developers",
-    "home.discord.description": "Vapor's 13k+ Discord community will be at your side to support you along the way. Ask questions, contribute and be a part of a thriving, wholesome corner of the internet.",
-    "home.discord.contribute": "Contribute",
-    // Home — API docs.
-    "home.apiDocs.title": "Get all the tools you need to build with Swift",
-    "home.apiDocs.goToApiDocs": "Go to API Docs",
-    "home.apiDocs.item1": "Easily create new projects with the Vapor Toolbox",
-    "home.apiDocs.item2": "Expansive documentation and API reference",
-    "home.apiDocs.item3": "Everything you need to build backends and APIs",
-    "home.apiDocs.item4": "Full support for Swift's concurrency model",
-    // Home — sponsors.
-    "home.sponsors.title": "Supported by our incredible sponsors and backers",
-    "home.sponsors.becomeSupporter": "Become a supporter",
-    "home.sponsors.prev": "Previous Sponsors",
-    "home.sponsors.next": "Next Sponsors",
-    // Showcase page chrome. (The app descriptions themselves stay English for now
-    // — the card data set is being made structured + localisable separately.)
-    "showcase.title": "Vapor powers thousands of high performing apps and sites",
-    "showcase.caption": "Discover the apps, sites and companies building on Vapor - and the teams behind them.",
-    "showcase.search": "Search",
-    "showcase.searchLabel": "Search showcase",
-    "showcase.sortLabel": "Sort showcase",
-    "showcase.sort.newest": "Newest first",
-    "showcase.sort.oldest": "Oldest first",
-    "showcase.sort.nameAsc": "Name (A–Z)",
-    "showcase.sort.nameDesc": "Name (Z–A)",
-    "showcase.empty": "No apps match your search.",
-    "showcase.paginationLabel": "Showcase pagination",
-    "showcase.cta.title": "Using Vapor in your app? We'd love to feature you!",
-    "showcase.cta.caption": "Help demonstrate all the different ways Vapor can be used.",
-    "showcase.cta.button": "Get in touch",
-]
-
-// German translations (best-effort — native-speaker review welcome). Any key not
-// listed here falls back to the English string above, so the build never shows a
-// raw key. Card/showcase/sponsor copy is intentionally left to fall back for now
-// (that content is being made structured + localisable separately).
-let germanStrings: [String: String] = [
-    "nav.closeMenu": "Menü schließen",
-    "nav.documentation": "Dokumentation",
-    "nav.frameworkDocs": "Framework-Dokumentation",
-    "nav.frameworkDocs.caption": "Erfahren Sie, wie Sie Vapor verwenden",
-    "nav.apiDocs": "API-Dokumentation",
-    "nav.apiDocs.caption": "Referenzdokumentation für Vapor",
-    "nav.selectTheme": "Thema auswählen",
-    "footer.joinDiscord": "Tritt unserem Discord bei",
-    "footer.supporters": "Unterstützer",
-    "comingSoon.message": "Diese Seite befindet sich noch im Aufbau – bleib dran für Updates!",
-    "pages.showcase.title": "Showcase",
-    "pages.team.title": "Team",
-    "pages.supporters.title": "Unterstützer",
-    "pages.evangelists.title": "Evangelisten",
-    "cta.getStarted": "Loslegen",
-    "cta.learnMore": "Mehr erfahren",
-    "cta.seeInAction": "In Aktion sehen",
-    "home.hero.title": "Swift, aber auf dem Server",
-    "home.hero.caption": "Vapor bietet eine sichere, performante und einfach zu bedienende Grundlage, um HTTP-Server, Backends und APIs in Swift zu erstellen",
-    "home.hero.stars": "{count} Sterne auf GitHub",
-    "home.hero.starsLoading": "… Sterne auf GitHub",
-    "home.companies.caption": "Im Einsatz bei Unternehmen wie:",
-    "home.packages.title": "Entwickle serverseitig mit den Werkzeugen, die du bereits kennst",
-    "home.packages.caption": "Du musst keine Sprache von Grund auf neu lernen (oder ein eigenes Team aufbauen) nur für dein Backend. Vapor baut auf Apples SwiftNIO auf, sodass du die Sprache verwenden kannst, die du bereits kennst und liebst.",
-    "home.packages.vapor.description": "Erstelle effiziente APIs in einer Sprache, die du liebst. Definiere Routen, sende und empfange JSON und baue HTTP-Server.",
-    "home.packages.fluent.description": "Erstelle Models und interagiere mit deiner Datenbank in nativem, sicherem Swift-Code, ohne SQL schreiben zu müssen.",
-    "home.packages.jwt.description": "Erstelle, signiere und verifiziere JSON Web Tokens in Swift. Basiert auf Swift Crypto.",
-    "home.packages.leaf.description": "Eine in Swift geschriebene Template-Engine. Generiere HTML für Web-Apps und E-Mails mit einer einfachen Syntax, die jeder nutzen kann.",
-    "home.features.performance.title": "Hochperformante APIs und Server",
-    "home.features.performance.description": "Dank einer nicht-blockierenden, ereignisgesteuerten Architektur kannst du mit Vapor hochperformante, skalierbare APIs und HTTP-Server erstellen. Mit dem Concurrency-Modell von Swift schreibst du klaren, wartbaren Code, der effizient und leicht zu lesen ist.",
-    "home.features.confidence.title": "Veröffentliche mit Zuversicht – sogar freitags",
-    "home.features.confidence.description": "Dank Vapors ausdrucksstarkem, protokollorientiertem Design kannst du deinen Code mit gutem Gewissen veröffentlichen. Durch unseren starken Fokus auf Typsicherheit werden viele Fehler und Probleme bereits früh vom Compiler erkannt.",
-    "home.integrations.label": "Integrationen",
-    "home.integrations.title": "Vollständige Integration in deinen gesamten Stack",
-    "home.integrations.description": "Vapors ausgereiftes Ökosystem umfasst über hundert offizielle und von der Community gepflegte, serverorientierte Swift-Pakete, die das Entwickeln deiner Anwendungen einfach und effizient machen.",
-    "home.showcase.title": "Verwendet von den Teams hinter diesen großartigen Apps",
-    "home.discord.title": "Tritt der größten Community von Swift-Backend-Entwicklern bei",
-    "home.discord.description": "Vapors über 13.000 Mitglieder zählende Discord-Community steht dir auf deinem Weg zur Seite. Stelle Fragen, bring dich ein und werde Teil einer lebendigen, freundlichen Ecke des Internets.",
-    "home.apiDocs.title": "Erhalte alle Werkzeuge, die du zum Entwickeln mit Swift brauchst",
-    "home.apiDocs.item1": "Erstelle mühelos neue Projekte mit der Vapor Toolbox",
-    "home.apiDocs.item2": "Umfangreiche Dokumentation und API-Referenz",
-    "home.apiDocs.item3": "Alles, was du zum Entwickeln von Backends und APIs brauchst",
-    "home.apiDocs.item4": "Volle Unterstützung für das Concurrency-Modell von Swift",
-    "home.sponsors.title": "Unterstützt von unseren großartigen Sponsoren und Förderern",
-    "home.sponsors.becomeSupporter": "Werde Unterstützer",
-    "home.apiDocs.goToApiDocs": "Zur API-Dokumentation",
-    "home.discord.contribute": "Mitwirken",
-    "home.showcase.seeFullList": "Vollständige Liste ansehen",
-    "home.showcase.prev": "Vorheriges Showcase",
-    "home.showcase.next": "Nächstes Showcase",
-    "home.sponsors.prev": "Vorherige Sponsoren",
-    "home.sponsors.next": "Nächste Sponsoren",
-    "showcase.title": "Vapor treibt Tausende leistungsstarke Apps und Websites an",
-    "showcase.caption": "Entdecke die Apps, Websites und Unternehmen, die auf Vapor setzen – und die Teams dahinter.",
-    "showcase.search": "Suchen",
-    "showcase.searchLabel": "Showcase durchsuchen",
-    "showcase.sortLabel": "Showcase sortieren",
-    "showcase.sort.newest": "Neueste zuerst",
-    "showcase.sort.oldest": "Älteste zuerst",
-    "showcase.sort.nameAsc": "Name (A–Z)",
-    "showcase.sort.nameDesc": "Name (Z–A)",
-    "showcase.empty": "Keine Apps entsprechen deiner Suche.",
-    "showcase.paginationLabel": "Showcase-Seitennavigation",
-    "showcase.cta.title": "Nutzt du Vapor in deiner App? Wir würden dich gerne vorstellen!",
-    "showcase.cta.caption": "Hilf zu zeigen, auf wie viele verschiedene Arten Vapor eingesetzt werden kann.",
-    "showcase.cta.button": "Kontakt aufnehmen",
-]
-
-// Remaining languages — translated incrementally. Any key a language omits falls
-// back to the English string above. Best-effort translations; native-speaker
-// review welcome.
-let spanishStrings: [String: String] = [
-    "home.hero.title": "Swift, pero en el servidor",
-    "home.hero.caption": "Vapor proporciona una base segura, eficiente y fácil de usar para crear servidores HTTP, backends y APIs en Swift",
-    // Buttons / controls.
-    "cta.getStarted": "Comenzar",
-    "cta.learnMore": "Más información",
-    "cta.seeInAction": "Míralo en acción",
-    "home.apiDocs.goToApiDocs": "Ir a la documentación de la API",
-    "home.discord.contribute": "Contribuir",
-    "home.showcase.seeFullList": "Ver lista completa",
-    "home.showcase.prev": "Showcase anterior",
-    "home.showcase.next": "Showcase siguiente",
-    "home.sponsors.becomeSupporter": "Conviértete en patrocinador",
-    "home.sponsors.prev": "Patrocinadores anteriores",
-    "home.sponsors.next": "Patrocinadores siguientes",
-    "showcase.title": "Vapor impulsa miles de apps y sitios de alto rendimiento",
-    "showcase.caption": "Descubre las apps, los sitios y las empresas que construyen sobre Vapor, y los equipos que están detrás.",
-    "showcase.search": "Buscar",
-    "showcase.searchLabel": "Buscar en los proyectos",
-    "showcase.sortLabel": "Ordenar los proyectos",
-    "showcase.sort.newest": "Más recientes primero",
-    "showcase.sort.oldest": "Más antiguos primero",
-    "showcase.sort.nameAsc": "Nombre (A–Z)",
-    "showcase.sort.nameDesc": "Nombre (Z–A)",
-    "showcase.empty": "Ningún proyecto coincide con tu búsqueda.",
-    "showcase.paginationLabel": "Paginación de proyectos",
-    "showcase.cta.title": "¿Usas Vapor en tu app? ¡Nos encantaría destacarla!",
-    "showcase.cta.caption": "Ayuda a demostrar las distintas formas en que se puede usar Vapor.",
-    "showcase.cta.button": "Ponte en contacto",
-    "footer.joinDiscord": "Únete a nuestro Discord",
-    // Nav / footer chrome.
-    "nav.documentation": "Documentación",
-    "nav.frameworkDocs": "Documentación del framework",
-    "nav.frameworkDocs.caption": "Aprende a usar Vapor",
-    "nav.apiDocs": "Documentación de la API",
-    "nav.apiDocs.caption": "Documentación de referencia de Vapor",
-    "nav.selectTheme": "Seleccionar tema",
-    "nav.closeMenu": "Cerrar el menú",
-    "footer.supporters": "Colaboradores",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} estrellas en GitHub",
-    "home.hero.starsLoading": "… estrellas en GitHub",
-    "home.companies.caption": "Impulsando a empresas como:",
-    "home.packages.title": "Desarrolla el lado del servidor con las herramientas que ya conoces",
-    "home.packages.caption": "No necesitas aprender un lenguaje desde cero (ni formar un equipo distinto) solo para tu backend. Vapor está construido sobre SwiftNIO de Apple, así que podrás usar el lenguaje que ya conoces y adoras.",
-    "home.packages.vapor.description": "Crea APIs eficientes en un lenguaje que te encanta. Define rutas, envía y recibe JSON y construye servidores HTTP.",
-    "home.packages.fluent.description": "Crea modelos e interactúa con tu base de datos en código Swift nativo y seguro, sin necesidad de escribir SQL.",
-    "home.packages.jwt.description": "Crea, firma y verifica JSON Web Tokens en Swift. Construido sobre Swift Crypto.",
-    "home.packages.leaf.description": "Un motor de plantillas escrito en Swift. Genera HTML para aplicaciones web y correos electrónicos con una sintaxis sencilla que cualquiera puede usar.",
-    "home.features.performance.title": "APIs y servidores de alto rendimiento",
-    "home.features.performance.description": "Construido con una arquitectura no bloqueante y orientada a eventos, Vapor te permite crear APIs y servidores HTTP escalables y de alto rendimiento. Gracias al modelo de concurrencia de Swift, puedes escribir código claro y mantenible que es eficiente y fácil de leer.",
-    "home.features.confidence.title": "Lanza con confianza, incluso los viernes",
-    "home.features.confidence.description": "Con el diseño expresivo y orientado a protocolos de Vapor, tendrás tranquilidad al lanzar tu código. Gracias a nuestro enfoque en la seguridad de tipos, el compilador detecta muchos errores y problemas desde el principio.",
-    "home.integrations.label": "Integraciones",
-    "home.integrations.title": "Integración total con todo tu stack",
-    "home.integrations.description": "El maduro ecosistema de Vapor incluye más de cien paquetes Swift orientados al servidor, oficiales y mantenidos por la comunidad, para que crear tus aplicaciones sea fácil y eficiente.",
-    "home.showcase.title": "Usado por los equipos detrás de estas increíbles apps",
-    "home.sponsors.title": "Con el apoyo de nuestros increíbles patrocinadores y colaboradores",
-    "home.discord.title": "Únete a la mayor comunidad de desarrolladores backend de Swift",
-    "home.discord.description": "La comunidad de Discord de Vapor, con más de 13.000 miembros, estará a tu lado para ayudarte en el camino. Haz preguntas, contribuye y forma parte de un rincón próspero y acogedor de internet.",
-    "home.apiDocs.title": "Obtén todas las herramientas que necesitas para desarrollar con Swift",
-    "home.apiDocs.item1": "Crea nuevos proyectos fácilmente con Vapor Toolbox",
-    "home.apiDocs.item2": "Amplia documentación y referencia de la API",
-    "home.apiDocs.item3": "Todo lo que necesitas para crear backends y APIs",
-    "home.apiDocs.item4": "Compatibilidad total con el modelo de concurrencia de Swift",
-    "comingSoon.message": "Esta página aún está en construcción. ¡Vuelve pronto para ver novedades!",
-    "pages.showcase.title": "Proyectos",
-    "pages.team.title": "Equipo",
-    "pages.supporters.title": "Colaboradores",
-    "pages.evangelists.title": "Evangelistas",
-]
-let frenchStrings: [String: String] = [
-    "home.hero.title": "Swift, mais sur le serveur",
-    "home.hero.caption": "Vapor fournit une base sûre, performante et facile à utiliser pour créer des serveurs HTTP, des backends et des API en Swift",
-    // Buttons / controls.
-    "cta.getStarted": "Commencer",
-    "cta.learnMore": "En savoir plus",
-    "cta.seeInAction": "Voir en action",
-    "home.apiDocs.goToApiDocs": "Accéder à la documentation de l’API",
-    "home.discord.contribute": "Contribuer",
-    "home.showcase.seeFullList": "Voir la liste complète",
-    "home.showcase.prev": "Vitrine précédente",
-    "home.showcase.next": "Vitrine suivante",
-    "home.sponsors.becomeSupporter": "Devenir soutien",
-    "home.sponsors.prev": "Sponsors précédents",
-    "home.sponsors.next": "Sponsors suivants",
-    "showcase.title": "Vapor propulse des milliers d'applications et de sites performants",
-    "showcase.caption": "Découvrez les applications, les sites et les entreprises qui s'appuient sur Vapor, et les équipes derrière eux.",
-    "showcase.search": "Rechercher",
-    "showcase.searchLabel": "Rechercher dans la vitrine",
-    "showcase.sortLabel": "Trier la vitrine",
-    "showcase.sort.newest": "Plus récents d'abord",
-    "showcase.sort.oldest": "Plus anciens d'abord",
-    "showcase.sort.nameAsc": "Nom (A–Z)",
-    "showcase.sort.nameDesc": "Nom (Z–A)",
-    "showcase.empty": "Aucune application ne correspond à votre recherche.",
-    "showcase.paginationLabel": "Pagination de la vitrine",
-    "showcase.cta.title": "Vous utilisez Vapor dans votre application ? Nous serions ravis de la mettre en avant !",
-    "showcase.cta.caption": "Aidez à montrer les différentes façons d'utiliser Vapor.",
-    "showcase.cta.button": "Nous contacter",
-    "footer.joinDiscord": "Rejoignez notre Discord",
-    // Nav / footer chrome.
-    "nav.documentation": "Documentation",
-    "nav.frameworkDocs": "Documentation du framework",
-    "nav.frameworkDocs.caption": "Apprenez à utiliser Vapor",
-    "nav.apiDocs": "Documentation de l’API",
-    "nav.apiDocs.caption": "Documentation de référence de Vapor",
-    "nav.selectTheme": "Sélectionner le thème",
-    "nav.closeMenu": "Fermer le menu",
-    "footer.supporters": "Soutiens",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} étoiles sur GitHub",
-    "home.hero.starsLoading": "… étoiles sur GitHub",
-    "home.companies.caption": "Utilisé par des entreprises comme :",
-    "home.packages.title": "Développez côté serveur avec les outils que vous maîtrisez déjà",
-    "home.packages.caption": "Pas besoin d’apprendre un langage à partir de zéro (ni de constituer une autre équipe) juste pour votre backend. Vapor repose sur SwiftNIO d’Apple, vous pourrez donc utiliser le langage que vous connaissez et appréciez déjà.",
-    "home.packages.vapor.description": "Créez des API efficaces dans un langage que vous aimez. Définissez des routes, envoyez et recevez du JSON et créez des serveurs HTTP.",
-    "home.packages.fluent.description": "Créez des modèles et interagissez avec votre base de données en code Swift natif et sûr, sans avoir à écrire de SQL.",
-    "home.packages.jwt.description": "Créez, signez et vérifiez des JSON Web Tokens en Swift. Basé sur Swift Crypto.",
-    "home.packages.leaf.description": "Un moteur de templates écrit en Swift. Générez du HTML pour vos applications web et vos e-mails avec une syntaxe simple, à la portée de tous.",
-    "home.features.performance.title": "Des API et des serveurs hautement performants",
-    "home.features.performance.description": "Bâti sur une architecture non bloquante et événementielle, Vapor vous permet de créer des API et des serveurs HTTP performants et évolutifs. Grâce au modèle de concurrence de Swift, vous écrivez un code clair et maintenable, efficace et facile à lire.",
-    "home.features.confidence.title": "Déployez en toute confiance, même le vendredi",
-    "home.features.confidence.description": "Grâce à la conception expressive et orientée protocole de Vapor, vous déployez votre code l’esprit tranquille. Avec notre fort accent sur la sûreté du typage, de nombreuses erreurs et problèmes sont détectés très tôt par le compilateur.",
-    "home.integrations.label": "Intégrations",
-    "home.integrations.title": "Une intégration complète avec l’ensemble de votre stack",
-    "home.integrations.description": "L’écosystème mature de Vapor comprend plus d’une centaine de paquets Swift orientés serveur, officiels et maintenus par la communauté, pour rendre le développement de vos applications simple et efficace.",
-    "home.showcase.title": "Utilisé par les équipes derrière ces applications incroyables",
-    "home.sponsors.title": "Soutenu par nos incroyables sponsors et donateurs",
-    "home.discord.title": "Rejoignez la plus grande communauté de développeurs backend Swift",
-    "home.discord.description": "La communauté Discord de Vapor, forte de plus de 13 000 membres, sera à vos côtés pour vous accompagner. Posez des questions, contribuez et faites partie d’un coin d’internet dynamique et bienveillant.",
-    "home.apiDocs.title": "Obtenez tous les outils dont vous avez besoin pour développer avec Swift",
-    "home.apiDocs.item1": "Créez facilement de nouveaux projets avec la Vapor Toolbox",
-    "home.apiDocs.item2": "Une documentation complète et une référence d’API",
-    "home.apiDocs.item3": "Tout ce dont vous avez besoin pour créer des backends et des API",
-    "home.apiDocs.item4": "Une prise en charge complète du modèle de concurrence de Swift",
-    "comingSoon.message": "Cette page est encore en construction, restez à l’écoute pour les mises à jour !",
-    "pages.showcase.title": "Vitrine",
-    "pages.team.title": "Équipe",
-    "pages.supporters.title": "Soutiens",
-    "pages.evangelists.title": "Évangélistes",
-]
-let italianStrings: [String: String] = [
-    "home.hero.title": "Swift, ma sul server",
-    "home.hero.caption": "Vapor offre una base sicura, performante e facile da usare per creare server HTTP, backend e API in Swift",
-    // Buttons / controls.
-    "cta.getStarted": "Inizia",
-    "cta.learnMore": "Scopri di più",
-    "cta.seeInAction": "Guardalo in azione",
-    "home.apiDocs.goToApiDocs": "Vai alla documentazione API",
-    "home.discord.contribute": "Contribuisci",
-    "home.showcase.seeFullList": "Vedi l’elenco completo",
-    "home.showcase.prev": "Showcase precedente",
-    "home.showcase.next": "Showcase successivo",
-    "home.sponsors.becomeSupporter": "Diventa sostenitore",
-    "home.sponsors.prev": "Sponsor precedenti",
-    "home.sponsors.next": "Sponsor successivi",
-    "showcase.title": "Vapor alimenta migliaia di app e siti ad alte prestazioni",
-    "showcase.caption": "Scopri le app, i siti e le aziende che si basano su Vapor, e i team dietro di essi.",
-    "showcase.search": "Cerca",
-    "showcase.searchLabel": "Cerca nella vetrina",
-    "showcase.sortLabel": "Ordina la vetrina",
-    "showcase.sort.newest": "Prima i più recenti",
-    "showcase.sort.oldest": "Prima i meno recenti",
-    "showcase.sort.nameAsc": "Nome (A–Z)",
-    "showcase.sort.nameDesc": "Nome (Z–A)",
-    "showcase.empty": "Nessuna app corrisponde alla tua ricerca.",
-    "showcase.paginationLabel": "Paginazione della vetrina",
-    "showcase.cta.title": "Usi Vapor nella tua app? Ci piacerebbe metterla in evidenza!",
-    "showcase.cta.caption": "Aiutaci a mostrare i tanti modi diversi in cui si può usare Vapor.",
-    "showcase.cta.button": "Mettiti in contatto",
-    "footer.joinDiscord": "Unisciti al nostro Discord",
-    // Nav / footer chrome.
-    "nav.documentation": "Documentazione",
-    "nav.frameworkDocs": "Documentazione del Framework",
-    "nav.frameworkDocs.caption": "Impara a usare Vapor",
-    "nav.apiDocs": "Documentazione API",
-    "nav.apiDocs.caption": "Documentazione di riferimento per Vapor",
-    "nav.selectTheme": "Seleziona il tema",
-    "nav.closeMenu": "Chiudi il menu",
-    "footer.supporters": "Sostenitori",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} stelle su GitHub",
-    "home.hero.starsLoading": "… stelle su GitHub",
-    "home.companies.caption": "Al servizio di aziende come:",
-    "home.packages.title": "Sviluppa lato server con gli strumenti che già conosci",
-    "home.packages.caption": "Non c’è bisogno di imparare un linguaggio da zero (o di mettere insieme un altro team) solo per il tuo backend. Vapor è costruito su SwiftNIO di Apple, così potrai usare il linguaggio che già conosci e ami.",
-    "home.packages.vapor.description": "Crea API efficienti in un linguaggio che ami. Definisci rotte, invia e ricevi JSON e costruisci server HTTP.",
-    "home.packages.fluent.description": "Crea modelli e interagisci con il tuo database in codice Swift nativo e sicuro, senza dover scrivere alcun SQL.",
-    "home.packages.jwt.description": "Crea, firma e verifica i JSON Web Token in Swift. Costruito su Swift Crypto.",
-    "home.packages.leaf.description": "Un motore di template scritto in Swift. Genera HTML per app web ed email con una sintassi semplice che chiunque può usare.",
-    "home.features.performance.title": "API e server ad alte prestazioni",
-    "home.features.performance.description": "Costruito con un’architettura non bloccante e orientata agli eventi, Vapor ti permette di creare API e server HTTP scalabili e ad alte prestazioni. Grazie al modello di concorrenza di Swift, puoi scrivere codice chiaro e manutenibile, efficiente e facile da leggere.",
-    "home.features.confidence.title": "Rilascia con sicurezza, anche di venerdì",
-    "home.features.confidence.description": "Grazie al design espressivo e orientato ai protocolli di Vapor, potrai rilasciare il tuo codice con serenità. Con la nostra forte attenzione alla type-safety, molti errori e problemi vengono individuati subito dal compilatore.",
-    "home.integrations.label": "Integrazioni",
-    "home.integrations.title": "Integrazione completa con il tuo intero stack",
-    "home.integrations.description": "Il maturo ecosistema di Vapor include oltre cento pacchetti Swift orientati al server, ufficiali e mantenuti dalla community, per rendere lo sviluppo delle tue applicazioni semplice ed efficiente.",
-    "home.showcase.title": "Usato dai team dietro queste fantastiche app",
-    "home.sponsors.title": "Sostenuto dai nostri incredibili sponsor e sostenitori",
-    "home.discord.title": "Unisciti alla più grande community di sviluppatori backend Swift",
-    "home.discord.description": "La community Discord di Vapor, con oltre 13.000 membri, sarà al tuo fianco per supportarti lungo il percorso. Fai domande, contribuisci e diventa parte di un angolo di internet vivace e accogliente.",
-    "home.apiDocs.title": "Ottieni tutti gli strumenti di cui hai bisogno per sviluppare con Swift",
-    "home.apiDocs.item1": "Crea facilmente nuovi progetti con la Vapor Toolbox",
-    "home.apiDocs.item2": "Documentazione completa e riferimento delle API",
-    "home.apiDocs.item3": "Tutto ciò di cui hai bisogno per creare backend e API",
-    "home.apiDocs.item4": "Supporto completo per il modello di concorrenza di Swift",
-    "comingSoon.message": "Questa pagina è ancora in costruzione, resta sintonizzato per gli aggiornamenti!",
-    "pages.showcase.title": "Vetrina",
-    "pages.team.title": "Team",
-    "pages.supporters.title": "Sostenitori",
-    "pages.evangelists.title": "Evangelisti",
-]
-let japaneseStrings: [String: String] = [
-    "home.hero.title": "Swift、ただしサーバー上で",
-    "home.hero.caption": "Vaporは、SwiftでHTTPサーバー、バックエンド、APIを構築するための安全で高性能、かつ使いやすい基盤を提供します",
-    // Buttons / controls.
-    "cta.getStarted": "始める",
-    "cta.learnMore": "詳しく見る",
-    "cta.seeInAction": "実際の動作を見る",
-    "home.apiDocs.goToApiDocs": "APIドキュメントへ",
-    "home.discord.contribute": "貢献する",
-    "home.showcase.seeFullList": "一覧を見る",
-    "home.showcase.prev": "前のショーケース",
-    "home.showcase.next": "次のショーケース",
-    "home.sponsors.becomeSupporter": "サポーターになる",
-    "home.sponsors.prev": "前のスポンサー",
-    "home.sponsors.next": "次のスポンサー",
-    "showcase.title": "Vapor は高性能な数千ものアプリやサイトを支えています",
-    "showcase.caption": "Vapor で構築されたアプリ、サイト、企業、そしてそれを支えるチームをご覧ください。",
-    "showcase.search": "検索",
-    "showcase.searchLabel": "ショーケースを検索",
-    "showcase.sortLabel": "ショーケースを並べ替え",
-    "showcase.sort.newest": "新しい順",
-    "showcase.sort.oldest": "古い順",
-    "showcase.sort.nameAsc": "名前（A–Z）",
-    "showcase.sort.nameDesc": "名前（Z–A）",
-    "showcase.empty": "検索に一致するアプリはありません。",
-    "showcase.paginationLabel": "ショーケースのページ送り",
-    "showcase.cta.title": "あなたのアプリで Vapor を使っていますか？ぜひご紹介させてください！",
-    "showcase.cta.caption": "Vapor のさまざまな活用方法を一緒に紹介しましょう。",
-    "showcase.cta.button": "お問い合わせ",
-    "footer.joinDiscord": "Discordに参加する",
-    // Nav / footer chrome.
-    "nav.documentation": "ドキュメント",
-    "nav.frameworkDocs": "フレームワークドキュメント",
-    "nav.frameworkDocs.caption": "Vaporの使い方を学ぶ",
-    "nav.apiDocs": "APIドキュメント",
-    "nav.apiDocs.caption": "Vaporのリファレンスドキュメント",
-    "nav.selectTheme": "テーマを選択",
-    "nav.closeMenu": "メニューを閉じる",
-    "footer.supporters": "サポーター",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "GitHubで{count}個のスター",
-    "home.hero.starsLoading": "GitHubで…個のスター",
-    "home.companies.caption": "次のような企業で利用されています:",
-    "home.packages.title": "すでに慣れ親しんだツールでサーバーサイドを構築",
-    "home.packages.caption": "バックエンドのためだけに、言語をゼロから学んだり（別のチームを編成したり）する必要はありません。VaporはAppleのSwiftNIO上に構築されているため、すでに知っていて愛用している言語をそのまま使えます。",
-    "home.packages.vapor.description": "好きな言語で効率的なAPIを構築。ルートを作成し、JSONを送受信し、HTTPサーバーを構築できます。",
-    "home.packages.fluent.description": "ネイティブで安全なSwiftコードでモデルを作成し、SQLを書くことなくデータベースを操作できます。",
-    "home.packages.jwt.description": "SwiftでJSON Web Tokenを作成、署名、検証。Swift Cryptoを基盤としています。",
-    "home.packages.leaf.description": "Swiftで書かれたテンプレートエンジン。誰でも使えるシンプルな構文で、WebアプリとメールのHTMLを生成できます。",
-    "home.features.performance.title": "高性能なAPIとサーバー",
-    "home.features.performance.description": "ノンブロッキングかつイベント駆動のアーキテクチャで構築されたVaporなら、高性能でスケーラブルなAPIとHTTPサーバーを構築できます。SwiftのConcurrencyモデルを使えば、効率的で読みやすく、明確で保守しやすいコードを書けます。",
-    "home.features.confidence.title": "金曜日でも自信を持ってリリース",
-    "home.features.confidence.description": "Vaporの表現力豊かでプロトコル指向の設計により、安心してコードをリリースできます。型安全性を重視しているため、多くのエラーや問題はコンパイラによって早い段階で検出されます。",
-    "home.integrations.label": "連携",
-    "home.integrations.title": "スタック全体とのフル連携",
-    "home.integrations.description": "Vaporの成熟したエコシステムには、公式およびコミュニティが管理する100を超えるサーバー向けSwiftパッケージが含まれており、アプリケーション開発を簡単かつ効率的にします。",
-    "home.showcase.title": "これらの素晴らしいアプリを手がけるチームが利用しています",
-    "home.sponsors.title": "素晴らしいスポンサーと支援者の皆様に支えられています",
-    "home.discord.title": "Swiftバックエンド開発者の最大のコミュニティに参加しよう",
-    "home.discord.description": "13,000人を超えるVaporのDiscordコミュニティが、あなたの歩みをサポートします。質問したり、貢献したりして、活気にあふれた居心地のよいインターネットの一角の一員になりましょう。",
-    "home.apiDocs.title": "Swiftでの開発に必要なすべてのツールを手に入れよう",
-    "home.apiDocs.item1": "Vapor Toolboxで新しいプロジェクトを簡単に作成",
-    "home.apiDocs.item2": "充実したドキュメントとAPIリファレンス",
-    "home.apiDocs.item3": "バックエンドとAPIの構築に必要なすべて",
-    "home.apiDocs.item4": "SwiftのConcurrencyモデルを完全サポート",
-    "comingSoon.message": "このページはまだ準備中です。最新情報をお待ちください！",
-    "pages.showcase.title": "ショーケース",
-    "pages.team.title": "チーム",
-    "pages.supporters.title": "サポーター",
-    "pages.evangelists.title": "エヴァンジェリスト",
-]
-let koreanStrings: [String: String] = [
-    "home.hero.title": "Swift, 하지만 서버에서",
-    "home.hero.caption": "Vapor는 Swift로 HTTP 서버, 백엔드 및 API를 구축할 수 있는 안전하고 성능이 뛰어나며 사용하기 쉬운 기반을 제공합니다",
-    // Buttons / controls.
-    "cta.getStarted": "시작하기",
-    "cta.learnMore": "자세히 보기",
-    "cta.seeInAction": "직접 확인하기",
-    "home.apiDocs.goToApiDocs": "API 문서로 이동",
-    "home.discord.contribute": "기여하기",
-    "home.showcase.seeFullList": "전체 목록 보기",
-    "home.showcase.prev": "이전 쇼케이스",
-    "home.showcase.next": "다음 쇼케이스",
-    "home.sponsors.becomeSupporter": "후원자 되기",
-    "home.sponsors.prev": "이전 후원자",
-    "home.sponsors.next": "다음 후원자",
-    "showcase.title": "Vapor는 수천 개의 고성능 앱과 사이트를 지원합니다",
-    "showcase.caption": "Vapor로 만들어진 앱, 사이트, 기업과 그 뒤의 팀들을 만나보세요.",
-    "showcase.search": "검색",
-    "showcase.searchLabel": "쇼케이스 검색",
-    "showcase.sortLabel": "쇼케이스 정렬",
-    "showcase.sort.newest": "최신순",
-    "showcase.sort.oldest": "오래된순",
-    "showcase.sort.nameAsc": "이름 (A–Z)",
-    "showcase.sort.nameDesc": "이름 (Z–A)",
-    "showcase.empty": "검색과 일치하는 앱이 없습니다.",
-    "showcase.paginationLabel": "쇼케이스 페이지 매김",
-    "showcase.cta.title": "앱에서 Vapor를 사용하고 계신가요? 소개해 드리고 싶어요!",
-    "showcase.cta.caption": "Vapor를 활용하는 다양한 방법을 함께 보여주세요.",
-    "showcase.cta.button": "문의하기",
-    "footer.joinDiscord": "Discord에 참여하기",
-    // Nav / footer chrome.
-    "nav.documentation": "문서",
-    "nav.frameworkDocs": "프레임워크 문서",
-    "nav.frameworkDocs.caption": "Vapor 사용법 알아보기",
-    "nav.apiDocs": "API 문서",
-    "nav.apiDocs.caption": "Vapor 레퍼런스 문서",
-    "nav.selectTheme": "테마 선택",
-    "nav.closeMenu": "메뉴 닫기",
-    "footer.supporters": "후원자",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "GitHub에서 {count}개의 스타",
-    "home.hero.starsLoading": "GitHub에서 …개의 스타",
-    "home.companies.caption": "다음과 같은 기업들이 사용합니다:",
-    "home.packages.title": "이미 알고 있는 도구로 서버 사이드를 구축하세요",
-    "home.packages.caption": "백엔드만을 위해 언어를 처음부터 배우거나 별도의 팀을 꾸릴 필요가 없습니다. Vapor는 Apple의 SwiftNIO를 기반으로 만들어져, 이미 알고 좋아하는 언어를 그대로 사용할 수 있습니다.",
-    "home.packages.vapor.description": "좋아하는 언어로 효율적인 API를 구축하세요. 라우트를 만들고 JSON을 주고받으며 HTTP 서버를 구축할 수 있습니다.",
-    "home.packages.fluent.description": "네이티브하고 안전한 Swift 코드로 모델을 만들고, SQL을 작성할 필요 없이 데이터베이스와 상호작용하세요.",
-    "home.packages.jwt.description": "Swift에서 JSON Web Token을 생성, 서명, 검증하세요. Swift Crypto를 기반으로 합니다.",
-    "home.packages.leaf.description": "Swift로 작성된 템플릿 엔진. 누구나 사용할 수 있는 간단한 문법으로 웹 앱과 이메일용 HTML을 생성하세요.",
-    "home.features.performance.title": "고성능 API와 서버",
-    "home.features.performance.description": "논블로킹 이벤트 기반 아키텍처로 구축된 Vapor를 사용하면 고성능의 확장 가능한 API와 HTTP 서버를 만들 수 있습니다. Swift의 동시성 모델을 활용해 효율적이고 읽기 쉬우며 명확하고 유지보수하기 좋은 코드를 작성할 수 있습니다.",
-    "home.features.confidence.title": "금요일에도 자신 있게 배포하세요",
-    "home.features.confidence.description": "Vapor의 표현력 있고 프로토콜 지향적인 설계 덕분에 안심하고 코드를 배포할 수 있습니다. 타입 안전성에 중점을 두어 많은 오류와 문제를 컴파일러가 일찍 잡아냅니다.",
-    "home.integrations.label": "통합",
-    "home.integrations.title": "전체 스택과의 완벽한 통합",
-    "home.integrations.description": "Vapor의 성숙한 생태계에는 공식 및 커뮤니티가 관리하는 100개가 넘는 서버 우선 Swift 패키지가 포함되어 있어, 애플리케이션 개발을 쉽고 효율적으로 만들어 줍니다.",
-    "home.showcase.title": "이 멋진 앱들을 만든 팀들이 사용합니다",
-    "home.sponsors.title": "멋진 후원자와 서포터 여러분의 지원을 받고 있습니다",
-    "home.discord.title": "최대 규모의 Swift 백엔드 개발자 커뮤니티에 참여하세요",
-    "home.discord.description": "13,000명이 넘는 Vapor의 Discord 커뮤니티가 여러분의 여정을 함께합니다. 질문하고, 기여하며, 활기차고 따뜻한 인터넷 공간의 일원이 되어 보세요.",
-    "home.apiDocs.title": "Swift로 개발하는 데 필요한 모든 도구를 만나보세요",
-    "home.apiDocs.item1": "Vapor Toolbox로 손쉽게 새 프로젝트를 만드세요",
-    "home.apiDocs.item2": "방대한 문서와 API 레퍼런스",
-    "home.apiDocs.item3": "백엔드와 API 구축에 필요한 모든 것",
-    "home.apiDocs.item4": "Swift의 동시성 모델 완벽 지원",
-    "comingSoon.message": "이 페이지는 아직 준비 중입니다. 업데이트를 기대해 주세요!",
-    "pages.showcase.title": "쇼케이스",
-    "pages.team.title": "팀",
-    "pages.supporters.title": "후원자",
-    "pages.evangelists.title": "에반젤리스트",
-]
-let dutchStrings: [String: String] = [
-    "home.hero.title": "Swift, maar dan op een server",
-    "home.hero.caption": "Vapor biedt een veilige, performante en gebruiksvriendelijke basis om HTTP-servers, backends en API's in Swift te bouwen",
-    // Buttons / controls.
-    "cta.getStarted": "Aan de slag",
-    "cta.learnMore": "Meer informatie",
-    "cta.seeInAction": "Zie het in actie",
-    "home.apiDocs.goToApiDocs": "Naar de API-documentatie",
-    "home.discord.contribute": "Bijdragen",
-    "home.showcase.seeFullList": "Volledige lijst bekijken",
-    "home.showcase.prev": "Vorige showcase",
-    "home.showcase.next": "Volgende showcase",
-    "home.sponsors.becomeSupporter": "Word ondersteuner",
-    "home.sponsors.prev": "Vorige sponsors",
-    "home.sponsors.next": "Volgende sponsors",
-    "showcase.title": "Vapor drijft duizenden goed presterende apps en sites aan",
-    "showcase.caption": "Ontdek de apps, sites en bedrijven die op Vapor bouwen – en de teams erachter.",
-    "showcase.search": "Zoeken",
-    "showcase.searchLabel": "Etalage doorzoeken",
-    "showcase.sortLabel": "Etalage sorteren",
-    "showcase.sort.newest": "Nieuwste eerst",
-    "showcase.sort.oldest": "Oudste eerst",
-    "showcase.sort.nameAsc": "Naam (A–Z)",
-    "showcase.sort.nameDesc": "Naam (Z–A)",
-    "showcase.empty": "Geen apps komen overeen met je zoekopdracht.",
-    "showcase.paginationLabel": "Etalagepaginering",
-    "showcase.cta.title": "Gebruik je Vapor in je app? We laten je graag zien!",
-    "showcase.cta.caption": "Help laten zien op hoeveel verschillende manieren Vapor gebruikt kan worden.",
-    "showcase.cta.button": "Neem contact op",
-    "footer.joinDiscord": "Word lid van onze Discord",
-    // Nav / footer chrome.
-    "nav.documentation": "Documentatie",
-    "nav.frameworkDocs": "Framework-documentatie",
-    "nav.frameworkDocs.caption": "Leer hoe je Vapor gebruikt",
-    "nav.apiDocs": "API-documentatie",
-    "nav.apiDocs.caption": "Referentiedocumentatie voor Vapor",
-    "nav.selectTheme": "Selecteer thema",
-    "nav.closeMenu": "Menu sluiten",
-    "footer.supporters": "Ondersteuners",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} sterren op GitHub",
-    "home.hero.starsLoading": "… sterren op GitHub",
-    "home.companies.caption": "Drijft bedrijven aan zoals:",
-    "home.packages.title": "Bouw server-side met de tools die je al kent",
-    "home.packages.caption": "Je hoeft geen taal vanaf nul te leren (of een apart team samen te stellen) alleen voor je backend. Vapor is gebouwd op Apple’s SwiftNIO, dus je kunt de taal gebruiken die je al kent en waardeert.",
-    "home.packages.vapor.description": "Bouw efficiënte API's in een taal die je geweldig vindt. Maak routes, verstuur en ontvang JSON en bouw HTTP-servers.",
-    "home.packages.fluent.description": "Maak modellen en werk met je database in native, veilige Swift-code zonder ook maar één regel SQL te schrijven.",
-    "home.packages.jwt.description": "Maak, onderteken en verifieer JSON Web Tokens in Swift. Gebouwd op Swift Crypto.",
-    "home.packages.leaf.description": "Een template-engine geschreven in Swift. Genereer HTML voor zowel webapps als e-mails met een eenvoudige syntaxis die iedereen kan gebruiken.",
-    "home.features.performance.title": "Hoogperformante API's en servers",
-    "home.features.performance.description": "Vapor is gebouwd op een non-blocking, event-driven architectuur en stelt je in staat hoogperformante, schaalbare API's en HTTP-servers te bouwen. Met het Concurrency-model van Swift schrijf je heldere, onderhoudbare code die efficiënt en gemakkelijk te lezen is.",
-    "home.features.confidence.title": "Lever met vertrouwen op, zelfs op vrijdag",
-    "home.features.confidence.description": "Met het expressieve, protocol-georiënteerde ontwerp van Vapor lever je je code met een gerust hart op. Door onze sterke focus op type-veiligheid worden veel fouten en problemen al vroeg door de compiler opgespoord.",
-    "home.integrations.label": "Integraties",
-    "home.integrations.title": "Volledige integratie met je hele stack",
-    "home.integrations.description": "Het volwassen ecosysteem van Vapor omvat meer dan honderd officiële en door de community onderhouden server-first Swift-packages om het bouwen van je applicaties eenvoudig en efficiënt te maken.",
-    "home.showcase.title": "Gebruikt door de teams achter deze geweldige apps",
-    "home.sponsors.title": "Ondersteund door onze geweldige sponsors en donateurs",
-    "home.discord.title": "Word lid van de grootste community van Swift-backendontwikkelaars",
-    "home.discord.description": "De Discord-community van Vapor met meer dan 13.000 leden staat je bij op je weg. Stel vragen, draag bij en word deel van een bloeiend, hartelijk hoekje van het internet.",
-    "home.apiDocs.title": "Krijg alle tools die je nodig hebt om met Swift te bouwen",
-    "home.apiDocs.item1": "Maak eenvoudig nieuwe projecten met de Vapor Toolbox",
-    "home.apiDocs.item2": "Uitgebreide documentatie en API-referentie",
-    "home.apiDocs.item3": "Alles wat je nodig hebt om backends en API's te bouwen",
-    "home.apiDocs.item4": "Volledige ondersteuning voor het concurrency-model van Swift",
-    "comingSoon.message": "Deze pagina is nog in aanbouw, houd updates in de gaten!",
-    "pages.showcase.title": "Etalage",
-    "pages.team.title": "Team",
-    "pages.supporters.title": "Ondersteuners",
-    "pages.evangelists.title": "Evangelisten",
-]
-let polishStrings: [String: String] = [
-    "home.hero.title": "Swift, ale na serwerze",
-    "home.hero.caption": "Vapor zapewnia bezpieczną, wydajną i łatwą w użyciu podstawę do tworzenia serwerów HTTP, backendów i interfejsów API w Swift",
-    // Buttons / controls.
-    "cta.getStarted": "Rozpocznij",
-    "cta.learnMore": "Dowiedz się więcej",
-    "cta.seeInAction": "Zobacz w działaniu",
-    "home.apiDocs.goToApiDocs": "Przejdź do dokumentacji API",
-    "home.discord.contribute": "Współtwórz",
-    "home.showcase.seeFullList": "Zobacz pełną listę",
-    "home.showcase.prev": "Poprzedni showcase",
-    "home.showcase.next": "Następny showcase",
-    "home.sponsors.becomeSupporter": "Zostań sponsorem",
-    "home.sponsors.prev": "Poprzedni sponsorzy",
-    "home.sponsors.next": "Następni sponsorzy",
-    "showcase.title": "Vapor napędza tysiące wydajnych aplikacji i witryn",
-    "showcase.caption": "Poznaj aplikacje, witryny i firmy oparte na Vaporze – oraz stojące za nimi zespoły.",
-    "showcase.search": "Szukaj",
-    "showcase.searchLabel": "Przeszukaj realizacje",
-    "showcase.sortLabel": "Sortuj realizacje",
-    "showcase.sort.newest": "Najnowsze najpierw",
-    "showcase.sort.oldest": "Najstarsze najpierw",
-    "showcase.sort.nameAsc": "Nazwa (A–Z)",
-    "showcase.sort.nameDesc": "Nazwa (Z–A)",
-    "showcase.empty": "Żadna aplikacja nie pasuje do wyszukiwania.",
-    "showcase.paginationLabel": "Paginacja realizacji",
-    "showcase.cta.title": "Używasz Vapora w swojej aplikacji? Chętnie ją wyróżnimy!",
-    "showcase.cta.caption": "Pomóż pokazać, na ile różnych sposobów można używać Vapora.",
-    "showcase.cta.button": "Skontaktuj się",
-    "footer.joinDiscord": "Dołącz do naszego Discorda",
-    // Nav / footer chrome.
-    "nav.documentation": "Dokumentacja",
-    "nav.frameworkDocs": "Dokumentacja frameworka",
-    "nav.frameworkDocs.caption": "Naucz się korzystać z Vapor",
-    "nav.apiDocs": "Dokumentacja API",
-    "nav.apiDocs.caption": "Dokumentacja referencyjna Vapor",
-    "nav.selectTheme": "Wybierz motyw",
-    "nav.closeMenu": "Zamknij menu",
-    "footer.supporters": "Wspierający",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} gwiazdek na GitHub",
-    "home.hero.starsLoading": "… gwiazdek na GitHub",
-    "home.companies.caption": "Napędza firmy takie jak:",
-    "home.packages.title": "Twórz po stronie serwera za pomocą narzędzi, które już znasz",
-    "home.packages.caption": "Nie musisz uczyć się języka od zera (ani budować osobnego zespołu) tylko dla swojego backendu. Vapor jest zbudowany na bazie SwiftNIO firmy Apple, więc możesz używać języka, który już znasz i lubisz.",
-    "home.packages.vapor.description": "Twórz wydajne API w języku, który kochasz. Definiuj trasy, wysyłaj i odbieraj JSON oraz buduj serwery HTTP.",
-    "home.packages.fluent.description": "Twórz modele i korzystaj z bazy danych w natywnym, bezpiecznym kodzie Swift, bez konieczności pisania SQL.",
-    "home.packages.jwt.description": "Twórz, podpisuj i weryfikuj tokeny JSON Web Token w Swift. Zbudowane na Swift Crypto.",
-    "home.packages.leaf.description": "Silnik szablonów napisany w Swift. Generuj HTML dla aplikacji webowych i e-maili dzięki prostej składni, z której skorzysta każdy.",
-    "home.features.performance.title": "Wydajne API i serwery",
-    "home.features.performance.description": "Zbudowany w oparciu o nieblokującą architekturę sterowaną zdarzeniami, Vapor pozwala tworzyć wydajne i skalowalne API oraz serwery HTTP. Dzięki modelowi współbieżności Swift możesz pisać przejrzysty, łatwy w utrzymaniu kod, który jest wydajny i czytelny.",
-    "home.features.confidence.title": "Wdrażaj z pewnością, nawet w piątki",
-    "home.features.confidence.description": "Dzięki ekspresyjnemu, zorientowanemu na protokoły projektowi Vapora wdrażasz swój kod ze spokojem. Dzięki naszemu silnemu naciskowi na bezpieczeństwo typów wiele błędów i problemów jest wychwytywanych wcześnie przez kompilator.",
-    "home.integrations.label": "Integracje",
-    "home.integrations.title": "Pełna integracja z całym Twoim stosem",
-    "home.integrations.description": "Dojrzały ekosystem Vapora obejmuje ponad sto oficjalnych i utrzymywanych przez społeczność pakietów Swift zorientowanych na serwer, dzięki którym tworzenie aplikacji jest łatwe i wydajne.",
-    "home.showcase.title": "Używany przez zespoły stojące za tymi wspaniałymi aplikacjami",
-    "home.sponsors.title": "Wspierany przez naszych niesamowitych sponsorów i darczyńców",
-    "home.discord.title": "Dołącz do największej społeczności programistów backendu w Swift",
-    "home.discord.description": "Licząca ponad 13 000 członków społeczność Discord Vapora będzie u Twojego boku, by wspierać Cię na każdym kroku. Zadawaj pytania, współtwórz i stań się częścią tętniącego życiem, przyjaznego zakątka internetu.",
-    "home.apiDocs.title": "Zdobądź wszystkie narzędzia potrzebne do tworzenia w Swift",
-    "home.apiDocs.item1": "Łatwo twórz nowe projekty dzięki Vapor Toolbox",
-    "home.apiDocs.item2": "Obszerna dokumentacja i dokumentacja API",
-    "home.apiDocs.item3": "Wszystko, czego potrzebujesz do tworzenia backendów i API",
-    "home.apiDocs.item4": "Pełne wsparcie dla modelu współbieżności Swift",
-    "comingSoon.message": "Ta strona jest jeszcze w budowie, śledź aktualizacje!",
-    "pages.showcase.title": "Realizacje",
-    "pages.team.title": "Zespół",
-    "pages.supporters.title": "Wspierający",
-    "pages.evangelists.title": "Ewangeliści",
-]
-let chineseStrings: [String: String] = [
-    "home.hero.title": "Swift，但在服务器上",
-    "home.hero.caption": "Vapor 提供了一个安全、高性能且易于使用的基础，用于在 Swift 中构建 HTTP 服务器、后端和 API",
-    // Buttons / controls.
-    "cta.getStarted": "开始使用",
-    "cta.learnMore": "了解更多",
-    "cta.seeInAction": "查看实际效果",
-    "home.apiDocs.goToApiDocs": "前往 API 文档",
-    "home.discord.contribute": "参与贡献",
-    "home.showcase.seeFullList": "查看完整列表",
-    "home.showcase.prev": "上一个展示",
-    "home.showcase.next": "下一个展示",
-    "home.sponsors.becomeSupporter": "成为赞助者",
-    "home.sponsors.prev": "上一批赞助者",
-    "home.sponsors.next": "下一批赞助者",
-    "showcase.title": "Vapor 为数千款高性能应用和网站提供支持",
-    "showcase.caption": "探索使用 Vapor 构建的应用、网站和公司，以及背后的团队。",
-    "showcase.search": "搜索",
-    "showcase.searchLabel": "搜索案例展示",
-    "showcase.sortLabel": "排序案例展示",
-    "showcase.sort.newest": "最新优先",
-    "showcase.sort.oldest": "最早优先",
-    "showcase.sort.nameAsc": "名称（A–Z）",
-    "showcase.sort.nameDesc": "名称（Z–A）",
-    "showcase.empty": "没有符合搜索条件的应用。",
-    "showcase.paginationLabel": "案例展示分页",
-    "showcase.cta.title": "在你的应用中使用了 Vapor 吗？我们很乐意展示它！",
-    "showcase.cta.caption": "帮助展示 Vapor 的各种使用方式。",
-    "showcase.cta.button": "联系我们",
-    "footer.joinDiscord": "加入我们的 Discord",
-    // Nav / footer chrome.
-    "nav.documentation": "文档",
-    "nav.frameworkDocs": "框架文档",
-    "nav.frameworkDocs.caption": "学习如何使用 Vapor",
-    "nav.apiDocs": "API 文档",
-    "nav.apiDocs.caption": "Vapor 参考文档",
-    "nav.selectTheme": "选择主题",
-    "nav.closeMenu": "关闭菜单",
-    "footer.supporters": "赞助者",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} 个 GitHub star",
-    "home.hero.starsLoading": "… 个 GitHub star",
-    "home.companies.caption": "为以下公司提供支持：",
-    "home.packages.title": "使用你已熟悉的工具构建服务器端",
-    "home.packages.caption": "无需仅仅为了后端而从头学习一门语言（或组建另一支团队）。Vapor 基于 Apple 的 SwiftNIO 构建，因此你可以继续使用你已经熟悉和喜爱的语言。",
-    "home.packages.vapor.description": "用你喜爱的语言构建高效的 API。创建路由、收发 JSON 并构建 HTTP 服务器。",
-    "home.packages.fluent.description": "使用原生、安全的 Swift 代码创建模型并操作数据库，无需编写任何 SQL。",
-    "home.packages.jwt.description": "在 Swift 中创建、签名和验证 JSON Web Token。基于 Swift Crypto 构建。",
-    "home.packages.leaf.description": "用 Swift 编写的模板引擎。使用任何人都能上手的简单语法，为 Web 应用和电子邮件生成 HTML。",
-    "home.features.performance.title": "高性能的 API 和服务器",
-    "home.features.performance.description": "Vapor 采用非阻塞、事件驱动的架构，让你能够构建高性能、可扩展的 API 和 HTTP 服务器。借助 Swift 的并发模型，你可以编写清晰、易维护、高效且易读的代码。",
-    "home.features.confidence.title": "即使在周五也能放心发布",
-    "home.features.confidence.description": "凭借 Vapor 富有表现力、面向协议的设计，你在发布代码时会更加安心。由于我们高度重视类型安全，许多错误和问题在编译阶段就能被及早发现。",
-    "home.integrations.label": "集成",
-    "home.integrations.title": "与你的整个技术栈完全集成",
-    "home.integrations.description": "Vapor 成熟的生态系统包含一百多个官方及社区维护的服务器优先 Swift 包，让你的应用开发更轻松、更高效。",
-    "home.showcase.title": "这些出色应用背后的团队都在使用",
-    "home.sponsors.title": "由我们了不起的赞助者和支持者提供支持",
-    "home.discord.title": "加入最大的 Swift 后端开发者社区",
-    "home.discord.description": "Vapor 拥有超过 13,000 名成员的 Discord 社区将一路为你提供支持。提出问题、参与贡献，成为这个充满活力、友善的网络角落的一员。",
-    "home.apiDocs.title": "获取使用 Swift 构建所需的全部工具",
-    "home.apiDocs.item1": "使用 Vapor Toolbox 轻松创建新项目",
-    "home.apiDocs.item2": "丰富的文档和 API 参考",
-    "home.apiDocs.item3": "构建后端和 API 所需的一切",
-    "home.apiDocs.item4": "全面支持 Swift 的并发模型",
-    "comingSoon.message": "此页面仍在建设中，敬请期待更新！",
-    "pages.showcase.title": "案例展示",
-    "pages.team.title": "团队",
-    "pages.supporters.title": "赞助者",
-    "pages.evangelists.title": "布道师",
-]
-let brazilianPortugueseStrings: [String: String] = [
-    "home.hero.title": "Swift, mas no servidor",
-    "home.hero.caption": "O Vapor oferece uma base segura, eficiente e fácil de usar para criar servidores HTTP, backends e APIs em Swift",
-    // Buttons / controls.
-    "cta.getStarted": "Começar",
-    "cta.learnMore": "Saiba mais",
-    "cta.seeInAction": "Veja em ação",
-    "home.apiDocs.goToApiDocs": "Ir para a documentação da API",
-    "home.discord.contribute": "Contribuir",
-    "home.showcase.seeFullList": "Ver lista completa",
-    "home.showcase.prev": "Showcase anterior",
-    "home.showcase.next": "Próximo showcase",
-    "home.sponsors.becomeSupporter": "Torne-se um apoiador",
-    "home.sponsors.prev": "Patrocinadores anteriores",
-    "home.sponsors.next": "Próximos patrocinadores",
-    "showcase.title": "O Vapor impulsiona milhares de apps e sites de alto desempenho",
-    "showcase.caption": "Conheça os apps, sites e empresas que constroem com o Vapor — e os times por trás deles.",
-    "showcase.search": "Buscar",
-    "showcase.searchLabel": "Buscar na vitrine",
-    "showcase.sortLabel": "Ordenar a vitrine",
-    "showcase.sort.newest": "Mais recentes primeiro",
-    "showcase.sort.oldest": "Mais antigos primeiro",
-    "showcase.sort.nameAsc": "Nome (A–Z)",
-    "showcase.sort.nameDesc": "Nome (Z–A)",
-    "showcase.empty": "Nenhum app corresponde à sua busca.",
-    "showcase.paginationLabel": "Paginação da vitrine",
-    "showcase.cta.title": "Usa o Vapor no seu app? Adoraríamos destacá-lo!",
-    "showcase.cta.caption": "Ajude a mostrar as diversas formas de usar o Vapor.",
-    "showcase.cta.button": "Entre em contato",
-    "footer.joinDiscord": "Entre no nosso Discord",
-    // Nav / footer chrome.
-    "nav.documentation": "Documentação",
-    "nav.frameworkDocs": "Documentação do framework",
-    "nav.frameworkDocs.caption": "Aprenda a usar o Vapor",
-    "nav.apiDocs": "Documentação da API",
-    "nav.apiDocs.caption": "Documentação de referência do Vapor",
-    "nav.selectTheme": "Selecionar tema",
-    "nav.closeMenu": "Fechar menu",
-    "footer.supporters": "Apoiadores",
-    // Home — stars / companies / packages.
-    "home.hero.stars": "{count} estrelas no GitHub",
-    "home.hero.starsLoading": "… estrelas no GitHub",
-    "home.companies.caption": "Impulsionando empresas como:",
-    "home.packages.title": "Crie o lado do servidor com as ferramentas que você já conhece",
-    "home.packages.caption": "Não é preciso aprender uma linguagem do zero (nem montar uma equipe separada) só para o seu backend. O Vapor é construído sobre o SwiftNIO da Apple, então você poderá usar a linguagem que você já conhece e ama.",
-    "home.packages.vapor.description": "Crie APIs eficientes em uma linguagem que você ama. Defina rotas, envie e receba JSON e construa servidores HTTP.",
-    "home.packages.fluent.description": "Crie modelos e interaja com seu banco de dados em código Swift nativo e seguro, sem precisar escrever nenhum SQL.",
-    "home.packages.jwt.description": "Crie, assine e verifique JSON Web Tokens em Swift. Construído sobre o Swift Crypto.",
-    "home.packages.leaf.description": "Um motor de templates escrito em Swift. Gere HTML para aplicações web e e-mails com uma sintaxe simples que qualquer pessoa pode usar.",
-    "home.features.performance.title": "APIs e servidores de alto desempenho",
-    "home.features.performance.description": "Construído com uma arquitetura não bloqueante e orientada a eventos, o Vapor permite criar APIs e servidores HTTP escaláveis e de alto desempenho. Com o modelo de concorrência do Swift, você pode escrever um código claro e fácil de manter, eficiente e legível.",
-    "home.features.confidence.title": "Faça deploy com confiança, mesmo às sextas-feiras",
-    "home.features.confidence.description": "Com o design expressivo e orientado a protocolos do Vapor, você terá tranquilidade ao publicar seu código. Com nosso forte foco em segurança de tipos, muitos erros e problemas são detectados cedo pelo compilador.",
-    "home.integrations.label": "Integrações",
-    "home.integrations.title": "Integração completa com toda a sua stack",
-    "home.integrations.description": "O ecossistema maduro do Vapor inclui mais de cem pacotes Swift voltados para o servidor, oficiais e mantidos pela comunidade, para tornar o desenvolvimento das suas aplicações fácil e eficiente.",
-    "home.showcase.title": "Usado pelas equipes por trás destes aplicativos incríveis",
-    "home.sponsors.title": "Apoiado por nossos incríveis patrocinadores e colaboradores",
-    "home.discord.title": "Junte-se à maior comunidade de desenvolvedores backend de Swift",
-    "home.discord.description": "A comunidade do Discord do Vapor, com mais de 13 mil membros, estará ao seu lado para ajudar você ao longo do caminho. Faça perguntas, contribua e faça parte de um cantinho próspero e acolhedor da internet.",
-    "home.apiDocs.title": "Obtenha todas as ferramentas necessárias para desenvolver com Swift",
-    "home.apiDocs.item1": "Crie novos projetos facilmente com o Vapor Toolbox",
-    "home.apiDocs.item2": "Documentação abrangente e referência da API",
-    "home.apiDocs.item3": "Tudo o que você precisa para criar backends e APIs",
-    "home.apiDocs.item4": "Suporte completo ao modelo de concorrência do Swift",
-    "comingSoon.message": "Esta página ainda está em construção. Fique de olho nas novidades!",
-    "pages.showcase.title": "Vitrine",
-    "pages.team.title": "Equipe",
-    "pages.supporters.title": "Apoiadores",
-    "pages.evangelists.title": "Evangelistas",
-]
+// Localisation: every `#localise("key")` resolves against a per-language
+// `customStrings` dictionary. Each language's strings + its `Language` value live
+// in `Translations/<Language>.swift`; English is the default, so any key a
+// translation omits falls back to English.
 
 // MARK: - Site
 
@@ -853,270 +42,17 @@ let site = KilnSite(
         palette: .autoLightDark(primary: .black, accent: .blue)
     ),
     languages: [
-        .init(.english, isDefault: true, customStrings: englishStrings, image: "static/images/opengraph/vapor-og-en-2x.png"),
-        .init(
-            .german,
-            customStrings: germanStrings,
-            image: "static/images/opengraph/vapor-og-de-2x.png",
-            localisation: .init(
-                notFoundTitle: "Seite nicht gefunden",
-                notFoundMessage: "Die gesuchte Seite wurde möglicherweise verschoben, umbenannt oder existiert nicht.",
-                notFoundLink: "Zurück zur Startseite",
-                toggleNavigation: "Navigation umschalten",
-                skipToContent: "Zum Inhalt springen",
-                home: "Startseite",
-                store: "Shop",
-                blog: "Blog",
-                showcase: "Showcase",
-                team: "Team",
-                help: "Hilfe",
-                pressKit: "Pressekit",
-                community: "Community",
-                resources: "Ressourcen",
-                language: "Sprache",
-                theme: "Thema",
-                lightTheme: "Hell",
-                darkTheme: "Dunkel",
-                systemTheme: "System"
-            )
-        ),
-        // Remaining languages (the docs' set plus Brazilian Portuguese). Being
-        // translated incrementally — keys not yet provided fall back to English,
-        // so the `/<locale>/` pages build now with English copy where untranslated.
-        .init(
-            .spanish,
-            customStrings: spanishStrings,
-            image: "static/images/opengraph/vapor-og-es-2x.png",
-            localisation: .init(
-                notFoundTitle: "Página no encontrada",
-                notFoundMessage: "Es posible que la página que buscas se haya movido, cambiado de nombre o que nunca haya existido.",
-                notFoundLink: "Volver a la página de inicio",
-                toggleNavigation: "Alternar la navegación",
-                skipToContent: "Saltar al contenido",
-                home: "Inicio",
-                store: "Tienda",
-                blog: "Blog",
-                showcase: "Proyectos",
-                team: "Equipo",
-                help: "Ayuda",
-                pressKit: "Kit de prensa",
-                community: "Comunidad",
-                resources: "Recursos",
-                language: "Idioma",
-                theme: "Tema",
-                lightTheme: "Claro",
-                darkTheme: "Oscuro",
-                systemTheme: "Sistema"
-            )
-        ),
-        .init(
-            .french,
-            customStrings: frenchStrings,
-            image: "static/images/opengraph/vapor-og-fr-2x.png",
-            localisation: .init(
-                notFoundTitle: "Page introuvable",
-                notFoundMessage: "La page que vous recherchez a peut-être été déplacée, renommée ou n’a jamais existé.",
-                notFoundLink: "Retour à la page d’accueil",
-                toggleNavigation: "Afficher/masquer la navigation",
-                skipToContent: "Passer au contenu",
-                home: "Accueil",
-                store: "Boutique",
-                blog: "Blog",
-                showcase: "Vitrine",
-                team: "Équipe",
-                help: "Aide",
-                pressKit: "Kit de presse",
-                community: "Communauté",
-                resources: "Ressources",
-                language: "Langue",
-                theme: "Thème",
-                lightTheme: "Clair",
-                darkTheme: "Sombre",
-                systemTheme: "Système"
-            )
-        ),
-        .init(
-            .italian,
-            customStrings: italianStrings,
-            image: "static/images/opengraph/vapor-og-it-2x.png",
-            localisation: .init(
-                notFoundTitle: "Pagina non trovata",
-                notFoundMessage: "La pagina che stai cercando potrebbe essere stata spostata, rinominata o non essere mai esistita.",
-                notFoundLink: "Torna alla home page",
-                toggleNavigation: "Attiva/disattiva la navigazione",
-                skipToContent: "Vai al contenuto",
-                home: "Home",
-                store: "Shop",
-                blog: "Blog",
-                showcase: "Vetrina",
-                team: "Team",
-                help: "Aiuto",
-                pressKit: "Kit stampa",
-                community: "Community",
-                resources: "Risorse",
-                language: "Lingua",
-                theme: "Tema",
-                lightTheme: "Chiaro",
-                darkTheme: "Scuro",
-                systemTheme: "Sistema"
-            )
-        ),
-        .init(
-            .japanese,
-            customStrings: japaneseStrings,
-            image: "static/images/opengraph/vapor-og-ja-2x.png",
-            localisation: .init(
-                notFoundTitle: "ページが見つかりません",
-                notFoundMessage: "お探しのページは移動または名称変更されたか、存在しない可能性があります。",
-                notFoundLink: "ホームページに戻る",
-                toggleNavigation: "ナビゲーションの切り替え",
-                skipToContent: "コンテンツにスキップ",
-                home: "ホーム",
-                store: "ストア",
-                blog: "ブログ",
-                showcase: "ショーケース",
-                team: "チーム",
-                help: "ヘルプ",
-                pressKit: "プレスキット",
-                community: "コミュニティ",
-                resources: "リソース",
-                language: "言語",
-                theme: "テーマ",
-                lightTheme: "ライト",
-                darkTheme: "ダーク",
-                systemTheme: "システム"
-            )
-        ),
-        .init(
-            .korean,
-            customStrings: koreanStrings,
-            image: "static/images/opengraph/vapor-og-ko-2x.png",
-            localisation: .init(
-                notFoundTitle: "페이지를 찾을 수 없습니다",
-                notFoundMessage: "찾고 있는 페이지가 이동되었거나 이름이 변경되었거나 존재하지 않을 수 있습니다.",
-                notFoundLink: "홈페이지로 돌아가기",
-                toggleNavigation: "내비게이션 전환",
-                skipToContent: "본문으로 건너뛰기",
-                home: "홈",
-                store: "스토어",
-                blog: "블로그",
-                showcase: "쇼케이스",
-                team: "팀",
-                help: "도움말",
-                pressKit: "프레스 킷",
-                community: "커뮤니티",
-                resources: "리소스",
-                language: "언어",
-                theme: "테마",
-                lightTheme: "라이트",
-                darkTheme: "다크",
-                systemTheme: "시스템"
-            )
-        ),
-        .init(
-            .dutch,
-            customStrings: dutchStrings,
-            image: "static/images/opengraph/vapor-og-nl-2x.png",
-            localisation: .init(
-                notFoundTitle: "Pagina niet gevonden",
-                notFoundMessage: "De pagina die je zoekt is mogelijk verplaatst, hernoemd of heeft nooit bestaan.",
-                notFoundLink: "Terug naar de startpagina",
-                toggleNavigation: "Navigatie schakelen",
-                skipToContent: "Naar de inhoud",
-                home: "Home",
-                store: "Winkel",
-                blog: "Blog",
-                showcase: "Etalage",
-                team: "Team",
-                help: "Help",
-                pressKit: "Perskit",
-                community: "Community",
-                resources: "Bronnen",
-                language: "Taal",
-                theme: "Thema",
-                lightTheme: "Licht",
-                darkTheme: "Donker",
-                systemTheme: "Systeem"
-            )
-        ),
-        .init(
-            .polish,
-            customStrings: polishStrings,
-            image: "static/images/opengraph/vapor-og-pl-2x.png",
-            localisation: .init(
-                notFoundTitle: "Nie znaleziono strony",
-                notFoundMessage: "Strona, której szukasz, mogła zostać przeniesiona, zmieniona lub nigdy nie istniała.",
-                notFoundLink: "Powrót do strony głównej",
-                toggleNavigation: "Przełącz nawigację",
-                skipToContent: "Przejdź do treści",
-                home: "Strona główna",
-                store: "Sklep",
-                blog: "Blog",
-                showcase: "Realizacje",
-                team: "Zespół",
-                help: "Pomoc",
-                pressKit: "Materiały prasowe",
-                community: "Społeczność",
-                resources: "Zasoby",
-                language: "Język",
-                theme: "Motyw",
-                lightTheme: "Jasny",
-                darkTheme: "Ciemny",
-                systemTheme: "Systemowy"
-            )
-        ),
-        .init(
-            .chinese,
-            customStrings: chineseStrings,
-            image: "static/images/opengraph/vapor-og-zh-2x.png",
-            localisation: .init(
-                notFoundTitle: "未找到页面",
-                notFoundMessage: "您要查找的页面可能已被移动、重命名或从未存在。",
-                notFoundLink: "返回首页",
-                toggleNavigation: "切换导航",
-                skipToContent: "跳到内容",
-                home: "首页",
-                store: "商店",
-                blog: "博客",
-                showcase: "案例展示",
-                team: "团队",
-                help: "帮助",
-                pressKit: "媒体资源",
-                community: "社区",
-                resources: "资源",
-                language: "语言",
-                theme: "主题",
-                lightTheme: "浅色",
-                darkTheme: "深色",
-                systemTheme: "跟随系统"
-            )
-        ),
-        .init(
-            .brazilianPortuguese,
-            customStrings: brazilianPortugueseStrings,
-            image: "static/images/opengraph/vapor-og-pt-BR-2x.png",
-            localisation: .init(
-                notFoundTitle: "Página não encontrada",
-                notFoundMessage: "A página que você procura pode ter sido movida, renomeada ou nunca ter existido.",
-                notFoundLink: "Voltar para a página inicial",
-                toggleNavigation: "Alternar navegação",
-                skipToContent: "Pular para o conteúdo",
-                home: "Início",
-                store: "Loja",
-                blog: "Blog",
-                showcase: "Vitrine",
-                team: "Equipe",
-                help: "Ajuda",
-                pressKit: "Kit de imprensa",
-                community: "Comunidade",
-                resources: "Recursos",
-                language: "Idioma",
-                theme: "Tema",
-                lightTheme: "Claro",
-                darkTheme: "Escuro",
-                systemTheme: "Sistema"
-            )
-        ),
+        englishLanguage,
+        germanLanguage,
+        spanishLanguage,
+        frenchLanguage,
+        italianLanguage,
+        japaneseLanguage,
+        koreanLanguage,
+        dutchLanguage,
+        polishLanguage,
+        chineseLanguage,
+        brazilianPortugueseLanguage,
     ]
 ) {
     Page("Home", "index.md")
